% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPOT.R
\name{build_K_matrix}
\alias{build_K_matrix}
\title{build_K_matrix}
\usage{
build_K_matrix(
  data,
  radii,
  use.K,
  homogeneous,
  marked,
  cell.type,
  K.diff,
  nlarge = 10000,
  print.progress = TRUE
)
}
\arguments{
\item{data}{The dataset of single cell spatial data. Must include column names
PID, id, x, y, and type if \code{marked}. PID corresponds to the patient IDs
which should be replicated across ROIs for each patient. id corresponds to the
image ID within a patient. x and y give the coordinates of each cell. type (if given)
is a cell type label.}

\item{radii}{Vector of length \eqn{P} giving the radii at which to evaluate the
spatial summary measure. The first value must be zero.}

\item{use.K}{Boolean. If \code{TRUE}, calculates Ripley's K for each radius and
sample. If \code{FALSE}, calculates Besag's L for each radius and sample.}

\item{homogeneous}{Boolean. If \code{TRUE}, calculates the homogeneous versions of
Ripley's K or Besag's L. If \code{FALSE}, calculates the inhomogeneous versions of
these summaries.}

\item{marked}{Boolean. If \code{TRUE}, subsets the data to consider only cells with the
labels given in the \code{cell.type} parameter. If \code{FALSE}, considers all detected cells
included in \code{data}.}

\item{cell.type}{The cell type labels to consider. Can be one string, e.g. \code{cell.type = "CD4 T cell"}, or
a vector of two cell types, e.g. \code{cell.type = c("CD4 T cell", "CD8 T cell")}. If \code{!marked}, leave
this parameter as \code{NULL}.}

\item{K.diff}{Boolean. If \code{TRUE}, returns the difference between the empirical and theoretical values of
Ripley's K or Besag's L. If \code{FALSE}, returns the empirical value of the spatial summary.}

\item{nlarge}{Parameter for the \code{Kest} and \code{Lest} function for efficiency. If there are more cells
than \code{nlarge}, will apply a border correction for edge effects instead of the isotropic edge correction.
Default is 10000.}

\item{print.progress}{Boolean. Should progress in calculating the spatial summary across samples be printed?
If \code{TRUE}, prints a progress bar.}
}
\value{
Returns two objects: \code{Kr} and \code{Kr.df}. \code{Kr} is a list of length \eqn{n} containing
the \eqn{P} spatial summary values at each radius. \code{Kr.df} is the data.frame containing the spatial summaries
in \code{Kr} but includes the PIDs and image ids.
}
\description{
Construct a matrix of dimension \eqn{n \times P} where \eqn{n} is the number
of samples and \eqn{P} is the number of radii under consideration of the
values of Ripley's K or Besag's L evaluated at each radius for each sample.
}
\details{
This function is primarily used within the \code{SPOT} function.
}
