% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCVModel.R
\name{buildCVModel}
\alias{buildCVModel}
\title{buildCVModel}
\usage{
buildCVModel(x, y, modellingFunction, control = list())
}
\arguments{
\item{x}{design matrix (sample locations)}

\item{y}{vector of observations at \code{x}}

\item{modellingFunction}{the model that shall be fitted to each data fold}

\item{control}{(list), with the options for the model building procedure:\cr}
}
\value{
set of models (class cvModel)
}
\description{
Build a set of models trained on different folds of cross-validated data.
Can be used to estimate the uncertainty of a given model type at any point.
}
