% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRandomSearchDummy.R
\name{buildrsdummy}
\alias{buildrsdummy}
\title{Build random search dummy model}
\usage{
buildrsdummy(x, y, control)
}
\arguments{
\item{x}{x (independent variables), not used.}

\item{y}{y (dependent variable), not used.}

\item{control}{control, not used.}
}
\description{
This function is used to emulate uniform random search with SPOT.
It is a placeholder for the surrogate model and
simply returns an empty list, with class "rsdummy".
}
