% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclone.R
\name{calculationMothes}
\alias{calculationMothes}
\title{Cyclone Simulation: Mothes}
\usage{
calculationMothes(x, cyclone, fluid)
}
\arguments{
\item{x}{a single particle diameter (scalar)}

\item{cyclone}{list of cyclone's geometrical parameters}

\item{fluid}{list of fluid parameters}
}
\value{
returns a vector with first element being the collection efficiency, second element being the pressure drop
}
\description{
Calculate cyclone collection efficiency and pressure drop according to Mothes.
}
\keyword{internal}
