% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funSring.R
\name{checkArrival}
\alias{checkArrival}
\title{checkArrival}
\usage{
checkArrival(probNewCustomer)
}
\arguments{
\item{probNewCustomer}{probability of an arrival of a new customer}
}
\value{
logical
}
\description{
Calculate arrival events for S-Ring
}
\examples{
checkArrival(0.5)

}
