% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotTools.R
\name{checkFeasibilityNlopGnIngres}
\alias{checkFeasibilityNlopGnIngres}
\title{Check feasibility for NLOPT_GN_ISRES}
\usage{
checkFeasibilityNlopGnIngres(x, control)
}
\arguments{
\item{x}{(1 x n)-matrix to be tested}

\item{control}{Control list for \code{\link{spot}} and \code{\link{spotLoop}}.

Generated with \code{\link{spotControl}}.}
}
\value{
logical (\code{TRUE} if feasible)
}
\description{
Returns \code{TRUE} if x does satisfy ineq constraint OR
no constraint function is used
}
