% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialInputCheck.R
\name{checkForNAs}
\alias{checkForNAs}
\title{Check for NAs in x lower and upper}
\usage{
checkForNAs(x, lower, upper)
}
\arguments{
\item{x}{is an optional start point (or set of start points), specified as a matrix. One row for each point, and one column for each optimized parameter.}

\item{lower}{is a vector that defines the lower boundary of search space. This determines also the dimensionality of the problem.}

\item{upper}{is a vector that defines the upper boundary of search space.}
}
\description{
Creates an error message if there are any nas given in x lower or upper
}
\keyword{internal}
