% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialInputCheck.R
\name{checkFunEvalsDesignSize}
\alias{checkFunEvalsDesignSize}
\title{Check funEvals Setting against designSize}
\usage{
checkFunEvalsDesignSize(x, lower, control, inSpotLoop)
}
\arguments{
\item{x}{is an optional start point (or set of start points), specified as a matrix. One row for each point, and one column for each optimized parameter.}

\item{lower}{is a vector that defines the lower boundary of search space. This determines also the dimensionality of the problem.}

\item{inSpotLoop}{Boolean indicating whether the check is called from within spotLoop or not}

\item{fun}{is the objective function. It should receive a matrix x and return a matrix y. In case the function uses external code and is noisy, an additional seed parameter may be used, see the \code{control$seedFun} argument below for details.}

\item{upper}{is a vector that defines the upper boundary of search space.}
}
\description{
Checks if the designSize will result in a larger value than funEvals. If so, return an error.
}
\keyword{internal}
