% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialInputCheck.R
\name{checkInputTypesInControl}
\alias{checkInputTypesInControl}
\title{Check input types in the spotControl list.}
\usage{
checkInputTypesInControl(control)
}
\arguments{
\item{control}{is a list with control settings for spot. See \code{\link{spotControl}}.}
}
\description{
specified variables are tested for their type in the control list. If a type mismatch is found, an error is thrown.
}
\keyword{internal}
