% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialInputCheck.R
\name{checkLowerNotEqualsUpper}
\alias{checkLowerNotEqualsUpper}
\title{Check That Lower and Upper are not Equal}
\usage{
checkLowerNotEqualsUpper(lower, upper)
}
\arguments{
\item{lower}{is a vector that defines the lower boundary of search space. This determines also the dimensionality of the problem.}

\item{upper}{is a vector that defines the upper boundary of search space.}
}
\description{
If any entries in lower and upper are equal, the parameter has no range and cant be optimized. 
In that case, an error is generated.
}
\keyword{internal}
