% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialInputCheck.R
\name{checkLowerSmallerThanUpper}
\alias{checkLowerSmallerThanUpper}
\title{Check That Lower is smaller than Upper}
\usage{
checkLowerSmallerThanUpper(lower, upper)
}
\arguments{
\item{lower}{is a vector that defines the lower boundary of search space. This determines also the dimensionality of the problem.}

\item{upper}{is a vector that defines the upper boundary of search space.}
}
\description{
Check if lower actually contains smaller values than upper. Otherwise a warning is generated.
}
\keyword{internal}
