% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialInputCheck.R
\name{checkTypesOfInput}
\alias{checkTypesOfInput}
\title{Check Input Types}
\usage{
checkTypesOfInput(x, lower, upper, control)
}
\arguments{
\item{x}{is an optional start point (or set of start points), specified as a matrix. One row for each point, and one column for each optimized parameter.}

\item{lower}{is a vector that defines the lower boundary of search space. This determines also the dimensionality of the problem.}

\item{upper}{is a vector that defines the upper boundary of search space.}

\item{control}{is a list with control settings for spot. See \code{\link{spotControl}}.}
}
\description{
Creates an error message if any of the input types are not numeric even though they are configured to be numeric
}
\keyword{internal}
