% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialInputCheck.R
\name{checkVerbosityLevels}
\alias{checkVerbosityLevels}
\title{Check correct verbosity levels}
\usage{
checkVerbosityLevels(control)
}
\arguments{
\item{control}{spot control list}
}
\description{
Create an error message if the given verbosity level is not allowed
}
\keyword{internal}
