% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotTools.R
\name{code2nat}
\alias{code2nat}
\title{Transform coded values to natural values}
\usage{
code2nat(x, a, b)
}
\arguments{
\item{x}{\code{matrix} of m n-dimensional input values
from the interval \code{[0;1]}, i.e, dim(x) = m x n}

\item{a}{\code{vector} of n-dimensional lower bound, i.e., length(a) = n}

\item{b}{\code{vector} of n-dimensional upper bound, i.e., length(b) = n}
}
\description{
Input values from the interval from zero to one, i.e., normalized
values,  are mapped to the interval from a to b.
}
\examples{
x <- matrix(runif(10),2)
a <- c(-1,1,2,3,4)
b <- c(1,2,3,4,5)
R <- code2nat(x,a,b)

}
