% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildKrigingDACE.R
\name{daceLikelihood}
\alias{daceLikelihood}
\title{Wrapper for Maximum Likelihood Estimation}
\usage{
daceLikelihood(theta, pars, nugget)
}
\arguments{
\item{theta}{model parameter vector to be evaluated}

\item{pars}{model option list, as created with \code{\link{dacePrepareFit}}}

\item{nugget}{Value for nugget. Default is -1, which means the nugget was optimized during MLE.}
}
\value{
the likelihood value as calculated by \code{\link{daceEvalFit}}
}
\description{
Returns the maximum likelihood for the model parameter optimization.
}
\seealso{
\code{\link{buildKrigingDACE}} \code{\link{daceEvalFit}}
}
\keyword{internal}
