% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRSM.R
\name{descentSpotRSM}
\alias{descentSpotRSM}
\title{Descent RSM model}
\usage{
descentSpotRSM(object)
}
\arguments{
\item{object}{RSM model (settings and parameters) of class \code{spotRSM}.}
}
\value{
list with
\describe{
	\item{\code{x}}{list of points along the path of steepest descent}
	\item{\code{y}}{corresponding predicted values}
}
}
\description{
Generate steps along the path of steepest descent for a RSM model.
This is only intended as a manual tool to use together with 
\code{\link{buildRSM}}.
}
\seealso{
\code{\link{buildRSM}}
}
