% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotTools.R
\name{diff0}
\alias{diff0}
\title{diff0}
\usage{
diff0(x)
}
\arguments{
\item{x}{input vector}
}
\value{
vector of differences
}
\description{
Calculate differences
}
\details{
Input vector length = output vector length
}
\examples{
x <- 1:10
diff0(x)
}
\concept{spotTools}
