% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/systemDependentParallel.R
\name{doParallel}
\alias{doParallel}
\title{Parallel execution of code, dependent on the operating system}
\usage{
doParallel(X, FUN, nCores = 2, ...)
}
\arguments{
\item{X}{vector with arguments to parallelize over}

\item{FUN}{function that shall be applied to each element of X}

\item{nCores}{\code{integer}. Defines the number of cores.}

\item{...}{optional arguments to FUN}
}
\description{
mclapply is only supported on linux and macOS. On Windows parlapply should be used.
This function switches between both dependent on the operating system of the user.
}
