% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotPredictCoForrester.R
\name{forrCoModel}
\alias{forrCoModel}
\title{Helper Function Forrester Co-Kriging}
\usage{
forrCoModel(fit, AX)
}
\arguments{
\item{fit}{co-Kriging fit of class \code{coforr}}

\item{AX}{differences between samples, as calculated in \code{forrCoBuilder}}
}
\value{
co-Kriging fit of class \code{coforr}, now finalized
}
\description{
Final calculations of the Co-Kriging fit
}
\references{
FORRESTER, A.I.J, SOBESTER A. & KEAN, A.J. (2007), Multi-Fidelity optimization via surrogate modelling. \emph{Proc. R. Soc. A} 463, 3251-3269. \cr
Forrester, Alexander I.J.; Sobester, Andras; Keane, Andy J. (2008). Engineering Design via Surrogate Modelling - A Practical Guide. John Wiley & Sons.
}
\keyword{internal}

