% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spotPredictCoForrester.R
\name{forrIF}
\alias{forrIF}
\title{helper function}
\usage{
forrIF(x, i)
}
\arguments{
\item{x}{object}

\item{i}{index}
}
\description{
helper function
}
\keyword{internal}

