% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotPredictForrester.R
\name{forrRegLikelihood}
\alias{forrRegLikelihood}
\title{Calculate negative log-likelihood}
\usage{
forrRegLikelihood(x, AX, Ay, opt.p = FALSE, uselambda = TRUE)
}
\arguments{
\item{x}{vector, containing log10(theta) and lambda}

\item{AX}{3 dimensional array, constructed by forrBuilder from the sample locations}

\item{Ay}{vector of observations at sample locations}
}
\value{
list with elements\cr
\code{NegLnLike}  concentrated log-likelihood *-1 for minimising \cr
\code{Psi} correlation matrix\cr
\code{Psinv} inverse of correlation matrix (to save computation time in predict.forr)\cr
\code{mu} \cr
\code{ssq}
}
\description{
Used to determine theta/lambda values for the Kriging model in \code{\link{forrBuilder}}.
}
\seealso{
\code{\link{spotPredictForrester}} \code{\link{forrBuilder}}
}
\keyword{internal}

