% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funContinuous.R
\name{funBard}
\alias{funBard}
\title{funBard}
\usage{
funBard(x)
}
\arguments{
\item{x}{matrix of points to evaluate with the function. 
Rows for points and columns for dimension.}
}
\value{
1-column matrix with resulting function values
}
\description{
The Bard Test Function
}
\details{
x0 = (1,1,1)
f = 8.21487...1e-3
f = 17.4286... at (0.8406..., -infty, -infty)
}
\examples{
x1 <- matrix(c(1,1),1,)
funBard(x1)

}
\references{
More, J. J., Garbow, B. S., and  Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\doi{10.1145/355934.355936}

BARD, Y. Comparison of gradient methods for the solution of nonlinear parameter estimation 
problems SIAM J. Numer. Anal. 7 (1970), 157-186.
}
