% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funContinuous.R
\name{funBeale}
\alias{funBeale}
\title{funBeale}
\usage{
funBeale(x)
}
\arguments{
\item{x}{matrix of points to evaluate with the function. 
Rows for points and columns for dimension.}
}
\value{
1-column matrix with resulting function values
}
\description{
Beale Test Function
}
\examples{
x1 <- matrix(c(1,1),1,)
funBeale(x1)

res <- spot(,funBeale,c(1,-1),c(5,2),control=list(funEvals=15))
plotModel(res$model)

}
\references{
Beale, E.M.L. On an interactive method of finding a local minimum of a function of more than
one variable. Tech. Rep. No. 25, Statistical Techniques Research Group, Princeton Univ., 
Princeton, N.J., 1958.

Rosenbrock, H. (1960).
An automatic method for finding the greatest or least value of a function.
\emph{The Computer Journal}, \emph{3}(3), 175-184.
\doi{10.1093/comjnl/3.3.175}
}
