% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funContinuous.R
\name{funBranin}
\alias{funBranin}
\title{funBranin (No. 1)}
\usage{
funBranin(x)
}
\arguments{
\item{x}{matrix of points to evaluate with the function. Rows for points and columns for dimension.}
}
\value{
1-column matrix with resulting function values
}
\description{
Branin Test Function
}
\examples{
x1 <- matrix(c(-pi, 12.275),1,)
funBranin(x1)

}
