% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funContinuous.R
\name{funBrownBs}
\alias{funBrownBs}
\title{funbrownBs (No. 10, More No. 4)}
\usage{
funBrownBs(x)
}
\arguments{
\item{x}{matrix of points to evaluate with the function.
Rows for points and columns for dimension.}
}
\value{
1-column matrix with resulting function values
}
\description{
2-dim Brown badly scaled Test Function
}
\details{
n=2, m=3
x0 = (1,1)
f=0 at (1e6, 2e-6)
}
\examples{
x1 <- matrix(c(1,1),1,)
funBrownBs(x1)

res <- spot(,fun=funBrownBs,c(-10,-10),c(10,10),control=list(funEvals=20))
plotModel(res$model, points = rbind(c(res$xbest[1], res$xbest[2]),c(1.098e-5,9.106)))

}
\references{
More, J. J., Garbow, B. S., and  Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\url{https://www.osti.gov/servlets/purl/6650344}
}
