% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funSring.R
\name{funCosts}
\alias{funCosts}
\title{funCosts}
\usage{
funCosts(x)
}
\arguments{
\item{x}{vector: weight multiplier \code{sigma} and number of elevators \code{ne}}
}
\value{
fitness (costs) as matrix
}
\description{
optimWrapper for getCosts
}
\details{
Evaluate synthetic cost function that is based on the number of waiting customers and 
the number elevators
}
\examples{
sigma = 1
ne = 10 
x <- matrix(c(sigma, ne), 1,)
funCosts(x)

}
