% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funContinuous.R
\name{funGauss}
\alias{funGauss}
\title{funGauss}
\usage{
funGauss(x)
}
\arguments{
\item{x}{matrix of points to evaluate with the function. 
Rows for points and columns for dimension.}
}
\value{
1-column matrix with resulting function values
}
\description{
Gaussian Test Function
}
\examples{
x1 <- matrix(c(1,1,1),1,)
funGauss(x1)

res1 <- spot(,funGauss, 
  c(-0.001,-0.007,-0.003),
  c(0.5,1.0,1.1),
  control=list(funEvals=15))
  plotModel(res1$model, which = 1:2)

}
\references{
Unpublished
}
