% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funContinuous.R
\name{funGoldsteinPrice}
\alias{funGoldsteinPrice}
\title{Goldstein-Price Test Function}
\usage{
funGoldsteinPrice(x)
}
\arguments{
\item{x}{(\code{m,2})-matrix of points to evaluate with the function.
Rows for points and columns for dimension.}
}
\value{
1-column matrix with resulting function values
}
\description{
An implementation of Booker et al.’s method on
a re-scaled/coded version of the 2-dim Goldstein–Price function
}
\examples{
x1 <- matrix(c(-pi, 12.275),1,)
funGoldsteinPrice(x1)

}
