% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funContinuous.R
\name{funJennSamp}
\alias{funJennSamp}
\title{funJennSamp (No. 12, More No 6)}
\usage{
funJennSamp(x)
}
\arguments{
\item{x}{matrix of points to evaluate with the function.
Rows for points and columns for dimension.}
}
\value{
1-column matrix with resulting function values
}
\description{
2-dim Jennrich and Sampson Function Test Function
}
\examples{
x1 <- matrix(c(1,1),1,)
funJennSamp(x1)

res <- spot(,funJennSamp,c(0,0),c(0.3,0.3))
plotModel(res$model)

}
\references{
More, J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\doi{10.1145/355934.355936}

Jennrich, R.I., and Sampson (1968).
Application of stepwise regression to nonlinear estimation.
\emph{Technometrics}, \emph{3}(3), 63-72.
\url{https://www.tandfonline.com/doi/abs/10.1080/00401706.1968.10490535}
}
