% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funMultiObjectiveOptim.R
\name{funMoo}
\alias{funMoo}
\title{funMoo}
\usage{
funMoo(x)
}
\arguments{
\item{x}{matrix of points to evaluate with the function. Rows for points and columns for dimension.}
}
\value{
n-column matrix with resulting function values
}
\description{
Multi-objective Test Function
}
\details{
funMultiObjectiveOptimization 

A multiobjective dummy testfunction
}
\examples{
x1 <- matrix(c(-pi, 12.275),1,)
funMoo(x1)
x <- matrix(c(1,1,2), ncol=3 )

}
