% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funNoise.R
\name{funNoise}
\alias{funNoise}
\title{funNoise}
\usage{
funNoise(x, fun = funSphere, mean = 0, sigma = 1)
}
\arguments{
\item{x}{input matrix of candidate solution}

\item{fun}{objective function. Default: \code{funSphere}}

\item{mean}{error mean. Default: 1}

\item{sigma}{error sd. Default: 1}
}
\value{
vector of noisy objective function values
}
\description{
noise function
}
\examples{
set.seed(123)
require(SPOT)
x <- matrix(1:10, 5,2)
funNoise(x)

}
