% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funOptimLecture.R
\name{funOptimLecture}
\alias{funOptimLecture}
\title{funOptimLecture}
\usage{
funOptimLecture(vec)
}
\arguments{
\item{vec}{input vector or matrix of candidate solution}
}
\value{
vector of objective function values
}
\description{
A testfunction used in the optimizaton lecture of the AIT Masters course at TH Koeln
}
