% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funContinuous.R
\name{funRosen2}
\alias{funRosen2}
\title{funRosen2 (No. 2a)}
\usage{
funRosen2(x)
}
\arguments{
\item{x}{matrix of points to evaluate with the function. Rows for points and columns for dimension.}
}
\value{
1-column matrix with resulting function values
}
\description{
Rosenbrock Test Function (2-dim)
}
\examples{
x1 <- matrix(c(-pi, 12.275),1,)
funRosen2(x1)

}
