% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funContinuous.R
\name{funShiftedSphere}
\alias{funShiftedSphere}
\title{funShiftedSphere (No. 4)}
\usage{
funShiftedSphere(x, a = 1)
}
\arguments{
\item{x}{matrix of points to evaluate with the function. Rows for points and columns for dimension.}

\item{a}{offset added, i.e., f = sum (x-a)^2. Default: \code{1}.}
}
\value{
1-column matrix with resulting function values
}
\description{
Shifted Sphere Test Function with optimum at x_opt = a and f(x_opt) = 0
}
\examples{
x1 <- matrix(c(-pi, 12.275),1,)
a <- 1
funShiftedSphere(x1, a)

}
\seealso{
\code{\link{funSphere}}
}
