% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funSring.R
\name{funSring}
\alias{funSring}
\title{funSring}
\usage{
funSring(x, opt = list(), ...)
}
\arguments{
\item{x}{perceptron weights}

\item{opt}{list of optional parameters, e.g., 
\describe{
    \item{\code{nElevators}}{number of elevators}
    \item{\code{probNewCustomer}}{probability pf a customer arrival}
    \item{\code{nIterations}}{Number of itertions}
    \item{\code{randomSeed}}{random seed}
  }}

\item{...}{additional parameters}
}
\value{
fitness (matrix with one column)
}
\description{
wrapper for \code{\link{sring}}
}
\examples{
set.seed(123)
numberStates = 200
sigma = 1
x = matrix( rnorm(n = 2*numberStates, 1, sigma), 1,)
funSring(x)

}
