% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulateKriging.R
\name{getCorrelationMatrix}
\alias{getCorrelationMatrix}
\title{Compute Correlation Matrix}
\usage{
getCorrelationMatrix(object, x)
}
\arguments{
\item{object}{fit of the Kriging model (settings and parameters), of class \code{kriging}.}

\item{x}{list of samples / data}
}
\value{
the correlation matrix
}
\description{
Compute the correlation matrix of samples x, given the model object.
}
\seealso{
\code{\link{simulate.kriging}}

\code{\link{predict.kriging}}
}
\keyword{internal}
