% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funSring.R
\name{getCosts}
\alias{getCosts}
\title{getCosts}
\usage{
getCosts(x, ...)
}
\arguments{
\item{x}{vector with \code{sigma} weight multiplier and \code{ne} number of elevators}

\item{...}{optional parameters passed to \code{funSring}}
}
\value{
fitness (costs)
}
\description{
Evaluate synthetic cost function that is based on the number of waiting customers and 
the number elevators
}
\details{
Note: To accelerate testing, nIterations was set to 1e3 (instead of 1e6)
}
\examples{
set.seed(123)
sigma = 1
ne = 10
x <- c(sigma, ne)
getCosts(x)

}
