% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotTools.R
\name{getMultiStartPoints}
\alias{getMultiStartPoints}
\title{Get Multi Start Points}
\usage{
getMultiStartPoints(x, y, control)
}
\arguments{
\item{x}{matrix of design points}

\item{y}{matrix of function values (f(x))}

\item{control}{Control list for \code{\link{spot}} and \code{\link{spotLoop}}.
Generated with \code{\link{spotControl}}.}
}
\value{
x0 matrix of restart points
}
\description{
Determine multi start points for optimization on the surrogate.
Combines the current best with additional random starting points for
optimization on the surrogate.
}
