% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotTools.R
\name{getPerformanceStats}
\alias{getPerformanceStats}
\title{get performance stats}
\usage{
getPerformanceStats(x, y)
}
\arguments{
\item{x}{matrix of n solutions (usually a (nxd)-matrix, where d is the problem dimension)}

\item{y}{matrix with objective values (usually a (nx1)-matrix}
}
\description{
determines mean performance
}
\details{
further stats will be added
}
\examples{
x <- matrix(1:10, ncol=2, byrow=TRUE)
y1 <- funSphere(x) +1
y2 <- funSphere(x) -1
x <- rbind(x,x)
y <- rbind(y1, y2)
M <- getPerformanceStats(x,y)

}
