% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infillGetFullPrediction.R
\name{infillGetFullPrediction}
\alias{infillGetFullPrediction}
\title{infillGetFullPrediction}
\usage{
infillGetFullPrediction(predictionList, model)
}
\arguments{
\item{predictionList}{The results of a predict.model call}

\item{model}{The surrogate model which was used for the prediction}
}
\value{
list, full model prediction
}
\description{
Wrapper to get the full model Prediction instead of just one target for plotting purposes.
}
\keyword{internal}
