% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialInputCheck.R
\name{initialInputCheck}
\alias{initialInputCheck}
\title{Initial Input Check of Spot Configuration}
\usage{
initialInputCheck(x = NULL, lower, upper, control = list(),
  inSpotLoop = F)
}
\arguments{
\item{x}{is an optional start point (or set of start points), specified as a matrix. One row for each point, and one column for each optimized parameter.}

\item{lower}{is a vector that defines the lower boundary of search space. This determines also the dimensionality of the problem.}

\item{upper}{is a vector that defines the upper boundary of search space.}

\item{control}{is a list with control settings for spot. See \code{\link{spotControl}}.}
}
\description{
This function takes the same inputs as given to the \code{spot} call and checks for
possible problems in the users configuration
}
\keyword{internal}
