% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCVModel.R
\name{linearAdaptedSE}
\alias{linearAdaptedSE}
\title{linearAdaptedSE}
\usage{
linearAdaptedSE(sOld, newdata, x)
}
\arguments{
\item{sOld}{numeric vector, old uncertainty values}

\item{newdata}{matrix, new data points for which the uncertainty is estimated}

\item{x}{matrix, already evaluated data points}
}
\value{
numeric vector, adapted uncertainty values
}
\description{
Linearly adapt the uncertainty estimation of a CV model regarding its distance to known neighbours
}
\keyword{internal}
