% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funContinuous.R
\name{makeSpotFunList}
\alias{makeSpotFunList}
\title{makeSpotFunList}
\usage{
makeSpotFunList(vector2Matrix = TRUE)
}
\arguments{
\item{vector2Matrix}{logical. Convert vector input to matrix.
Default: TRUE, so it can be used with \code{\link[stats]{optim}}.}
}
\value{
list of functions
}
\description{
Generate a list of spot functions
}
\examples{
fr <- makeSpotFunList()
optim(c(-1.2,1), fr[[1]])

}
