% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCVModel.R
\name{maxNearestNeighbourDistance}
\alias{maxNearestNeighbourDistance}
\title{maxNearestNeighbourDistance}
\usage{
maxNearestNeighbourDistance(x)
}
\arguments{
\item{x}{matrix with candidate solutions}
}
\value{
maximum euclidean distance between two nearest neighbours
}
\description{
Find the maximum distance between 2 nearest neighbours in a data set
}
