% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildKrigingForrester.R
\name{normalizeMatrix2}
\alias{normalizeMatrix2}
\title{Normalize design 2}
\usage{
normalizeMatrix2(x, ymin, ymax, xmin, xmax)
}
\arguments{
\item{x}{design matrix in input space (n rows for each point, k columns for each parameter)}

\item{ymin}{minimum vector of normalized space}

\item{ymax}{maximum vector of normalized space}

\item{xmin}{minimum vector of input space}

\item{xmax}{maximum vector of input space}
}
\value{
normalized design matrix
}
\description{
Normalize design with given maximum and minimum in input space. Supportive function for Kriging model, not to be used directly.
}
\seealso{
\code{\link{buildKriging}}
}
\keyword{internal}
