% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repeatsOCBA.R
\name{ocbaRanking}
\alias{ocbaRanking}
\title{ocbaRanking}
\usage{
ocbaRanking(x, y, fun, control, ...)
}
\arguments{
\item{x}{matrix of x values}

\item{y}{matrix of y values, one dimensional!}

\item{fun}{objective function}

\item{control}{control list, see \code{\link{spotControl}}}

\item{...}{additional arguments to fun}
}
\value{
(x,y) matrix of sorted (by y) values. In case of noise are these values aggregated (y-mean) values.
}
\description{
Return the ocba ranking (xbest, ybest) for noisy optimization
}
\details{
Based on \code{\link{repeatsOCBA}}
}
