% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimRanddomSearch.R
\name{optimRSfun}
\alias{optimRSfun}
\title{Random search surrogate-optimizer}
\usage{
optimRSfun(x, fun, lower, upper, control, ...)
}
\arguments{
\item{x}{start guess, not used.}

\item{fun}{objective function to be evaluated via random search.}

\item{lower}{bound on the independent variables (search space).}

\item{upper}{bound on the independent variables (search space).}

\item{control}{not used.}

\item{...}{additional arguments, not used.}
}
\value{
list
}
\description{
This function is used to emulate uniform random search with SPOT.
It is used as the optimizer that searches for new candidates.
It returns a single uniform random sample within
the given lower and upper bounds of the search space.
}
