% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funSring.R
\name{perceptron}
\alias{perceptron}
\title{perceptron}
\usage{
perceptron(currentState, nStates, sElevator, sCustomer, weightsPerceptron)
}
\arguments{
\item{currentState}{current state for decision (num)}

\item{nStates}{numer of states (int)}

\item{sElevator}{elevators vector (logical)}

\item{sCustomer}{customer vector (logical)}

\item{weightsPerceptron}{Weight vector (num)}
}
\value{
logical pass or take decision
}
\description{
Perceptron to calculate decisions
}
\details{
Number of weights in NN controller is 2xnStates, 
  for each state (sElevator/sCustomer) there is one input
}
