% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRSM.R
\name{plot.spotRSM}
\alias{plot.spotRSM}
\title{Plot RSM model}
\usage{
\method{plot}{spotRSM}(x, ...)
}
\arguments{
\item{x}{RSM model (settings and parameters) of class \code{spotRSM}.}

\item{...}{parameters passed to plotting function (\code{contour})}
}
\description{
Plot model produced by \code{\link{buildRSM}}.
}
\keyword{internal}
