% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTree.R
\name{plot.spotTreeModel}
\alias{plot.spotTreeModel}
\title{Plot rpart tree model}
\usage{
\method{plot}{spotTreeModel}(x, ...)
}
\arguments{
\item{x}{tree model (settings and parameters) of class \code{spotTreeModel}.}

\item{...}{parameters passed to rpart.plot plotting function}
}
\description{
Plot model produced by \code{\link{buildTreeModel}}.
}
\keyword{internal}
