% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResults.R
\name{plotBestObj}
\alias{plotBestObj}
\title{Plot Best Objective Value}
\usage{
plotBestObj(y, end = length(y))
}
\arguments{
\item{y}{result vector}

\item{end}{length. Default: \code{length(y)}}
}
\value{
plot
}
\description{
Plot Best Objective Value
}
