% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotSingleDimFunction.R
\name{plotSingleDimFunction}
\alias{plotSingleDimFunction}
\title{plotSingleDimFunction}
\usage{
plotSingleDimFunction(evalFun, lower, upper, target)
}
\arguments{
\item{evalFun}{}

\item{lower}{}

\item{upper}{}

\item{target}{}
}
\description{
Plot a single dimensional Function
}
