% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildCVModel.R
\name{predict.cvModel}
\alias{predict.cvModel}
\title{predict.cvModel}
\usage{
\method{predict}{cvModel}(object, newdata, ...)
}
\arguments{
\item{object}{Kriging model (settings and parameters) of class \code{kriging}.}

\item{newdata}{design matrix to be predicted}

\item{...}{Additional parameters passed to the model}
}
\value{
prediction results: list with predicted mean ('y'), estimated uncertainty ('y'), linearly adapted uncertainty ('sLinear')
}
\description{
Predict with the cross validated model
}
