% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildEnsembleStack.R
\name{predict.ensembleStack}
\alias{predict.ensembleStack}
\title{Predict Stacked Ensemble}
\usage{
\method{predict}{ensembleStack}(object, newdata, ...)
}
\arguments{
\item{object}{Kriging model (settings and parameters) of class \code{kriging}.}

\item{newdata}{design matrix to be predicted}

\item{...}{not used}
}
\value{
list with predicted value \code{y}
}
\description{
Predict with ensemble model produced by \code{\link{buildEnsembleStack}}.
}
\seealso{
\code{\link{buildKriging}}
}
\keyword{internal}
