% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRandomSearchDummy.R
\name{predict.rsdummy}
\alias{predict.rsdummy}
\title{Predict random search dummy}
\usage{
\method{predict}{rsdummy}(object, newdata, ...)
}
\arguments{
\item{object}{dummy object, not used.}

\item{newdata}{dummy data, not used.}

\item{...}{additional arguments, not used.}
}
\description{
This function is used to emulate uniform random search with SPOT.
It is a placeholder for the surrogate model and
simply returns \code{NA} when called.
}
\keyword{internal}
