% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildBO.R
\name{predict.spotBOModel}
\alias{predict.spotBOModel}
\title{Prediction method for bayesian optimization model}
\usage{
\method{predict}{spotBOModel}(object, newdata, ...)
}
\arguments{
\item{object}{fit of the model, an object of class \code{"spotBOModel"}, produced by \code{\link{buildBO}}.}

\item{newdata}{matrix of new data.}

\item{...}{not used}
}
\description{
Wrapper for \code{predict.spotBOModel}.
}
\keyword{internal}
