% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGaussianProcessModel.R
\name{predict.spotGaussianProcessModel}
\alias{predict.spotGaussianProcessModel}
\title{Prediction method for Gaussian Process Model}
\usage{
\method{predict}{spotGaussianProcessModel}(object, newdata, ...)
}
\arguments{
\item{object}{fit of the model, an object of class \code{"spotGaussianProcessModel"}, produced by \code{\link{buildGaussianProcess}}.}

\item{newdata}{matrix of new data.}

\item{...}{not used}
}
\description{
Prediction method for Gaussian Process Model
}
\keyword{internal}
