% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildLasso.R
\name{predict.spotLassoModel}
\alias{predict.spotLassoModel}
\title{Prediction method for lasso model}
\usage{
\method{predict}{spotLassoModel}(object, newdata, ...)
}
\arguments{
\item{object}{fit of the model, an object of class \code{"spotLassoModel"}, 
produced by \code{\link{buildLasso}}.}

\item{newdata}{matrix of new data.}

\item{...}{not used}
}
\description{
Wrapper for \code{predict.glmnet}.
}
\keyword{internal}
