% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildLM.R
\name{predict.spotLinearModel}
\alias{predict.spotLinearModel}
\title{Prediction method for linear model}
\usage{
\method{predict}{spotLinearModel}(object, newdata, ...)
}
\arguments{
\item{object}{fit of the model, an object of class \code{"spotLinearModel"}, produced by \code{\link{buildLM}}.}

\item{newdata}{matrix of new data.}

\item{...}{not used}
}
\description{
Wrapper for \code{predict.lm}.
}
\keyword{internal}
