% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildLOESS.R
\name{predict.spotLOESS}
\alias{predict.spotLOESS}
\title{Predict loess model}
\usage{
\method{predict}{spotLOESS}(object, newdata, ...)
}
\arguments{
\item{object}{loess model (settings and parameters) of class \code{spotLOESS}.}

\item{newdata}{design matrix to be predicted}

\item{...}{not used}
}
\value{
list with predicted value \code{y}, standard error 
\code{s} (if object$target contains \code{"s"}) and
\code{ei} (if object$target contains \code{"ei"}) and
}
\description{
Predict with model produced by \code{\link{buildLOESS}}.
}
\seealso{
\code{\link{buildLOESS}}
}
\keyword{internal}
