% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRandomForest.R
\name{predict.spotRandomForest}
\alias{predict.spotRandomForest}
\title{Prediction method for random forest}
\usage{
\method{predict}{spotRandomForest}(object, newdata, ...)
}
\arguments{
\item{object}{fit of the model, an object of class \code{"spotRandomForest"}, produced by \code{\link{buildRandomForest}}.}

\item{newdata}{matrix of new data.}

\item{...}{not used}
}
\description{
Wrapper for \code{predict.randomForest}.
}
\keyword{internal}
