% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRanger.R
\name{predict.spotRanger}
\alias{predict.spotRanger}
\title{Predictor for spotExtraTrees}
\usage{
\method{predict}{spotRanger}(object, newdata, ...)
}
\arguments{
\item{object}{fit of the model, an object of class \code{"spotRandomForest"}, produced by \code{\link{buildRandomForest}}.}

\item{newdata}{matrix of new data.}

\item{...}{not used}
}
\description{
Wrapper for \code{predict.ranger}.
}
\keyword{internal}
