% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRSM.R
\name{predict.spotRSM}
\alias{predict.spotRSM}
\title{Predict RSM model}
\usage{
\method{predict}{spotRSM}(object, newdata, ...)
}
\arguments{
\item{object}{RSM model (settings and parameters) of class \code{spotRSM}.}

\item{newdata}{design matrix to be predicted}

\item{...}{not used}
}
\value{
list with predicted value \code{y}
}
\description{
Predict with model produced by \code{\link{buildRSM}}.
}
\seealso{
\code{\link{buildRSM}}
}
\keyword{internal}
