% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSO.R
\name{predict.spotSecondOrderModel}
\alias{predict.spotSecondOrderModel}
\title{Prediction method for linear model}
\usage{
\method{predict}{spotSecondOrderModel}(object, newdata, ...)
}
\arguments{
\item{object}{fit of the model, an object of class \code{"spotSecondOrderModel"}, produced by \code{\link{buildSO}}.}

\item{newdata}{matrix of new data.}

\item{...}{not used}
}
\description{
Wrapper for \code{predict.lm}.
}
\keyword{internal}
