% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTree.R
\name{predict.spotTreeModel}
\alias{predict.spotTreeModel}
\title{Prediction method for rpart tree models}
\usage{
\method{predict}{spotTreeModel}(object, newdata, ...)
}
\arguments{
\item{object}{fit of the model, an object of class \code{"spotTreeModel"}, produced by \code{\link{buildTreeModel}}.}

\item{newdata}{matrix of new data.}

\item{...}{not used}
}
\description{
Wrapper for \code{predict.rpart}.
}
\keyword{internal}
