% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotResults.R
\name{prepareBestObjectiveVal}
\alias{prepareBestObjectiveVal}
\title{Preprocess y Values to Plot Best Objective Value}
\usage{
prepareBestObjectiveVal(y, end = length(y))
}
\arguments{
\item{y}{result vector}

\item{end}{length. Default: \code{length(y)}}
}
\value{
prog
}
\description{
Preprocess y Values to Plot Best Objective Value
}
