\name{print.spotConfig}
\alias{print.spotConfig}
\title{Print function for spotConfig class}
\usage{
  \method{print}{spotConfig} (x, ...)
}
\arguments{
  \item{x}{spotConfig}

  \item{...}{additional parameters}
}
\description{
  Print function to summarize a spotConfig.
}
\keyword{internal}

