% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildBO.R
\name{print.spotBOModel}
\alias{print.spotBOModel}
\title{Print method for BO model}
\usage{
\method{print}{spotBOModel}(x, ...)
}
\arguments{
\item{...}{not used}

\item{object}{fit of the model, an object of class \code{"spotBOModel"}, produced by \code{\link{buildBO}}.}
}
\description{
Wrapper for \code{print.spotBOModel}.
}
\keyword{internal}
