% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildGaussianProcessModel.R
\name{print.spotGaussianProcessModel}
\alias{print.spotGaussianProcessModel}
\title{Print method for Gaussian Process Model}
\usage{
\method{print}{spotGaussianProcessModel}(x, ...)
}
\arguments{
\item{...}{not used}

\item{object}{fit of the model, an object of class \code{"spotGaussianProcessModel"}, produced by \code{\link{buildGaussianProcess}}.}
}
\description{
Print method for Gaussian Process Model
}
\keyword{internal}
