% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildLasso.R
\name{print.spotLassoModel}
\alias{print.spotLassoModel}
\title{Print method for lasso model}
\usage{
\method{print}{spotLassoModel}(x, ...)
}
\arguments{
\item{...}{not used}

\item{object}{fit of the model, an object of class \code{"spotLassoModel"}, 
produced by \code{\link{buildLasso}}.}
}
\description{
Wrapper for \code{print.lasso}.
}
\keyword{internal}
