% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildLM.R
\name{print.spotLinearModel}
\alias{print.spotLinearModel}
\title{Print method for linear model}
\usage{
\method{print}{spotLinearModel}(x, ...)
}
\arguments{
\item{...}{not used}

\item{object}{fit of the model, an object of class \code{"spotLinearModel"}, produced by \code{\link{buildLM}}.}
}
\description{
Wrapper for \code{print.lm}.
}
\keyword{internal}
