% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRandomForest.R
\name{print.spotRandomForest}
\alias{print.spotRandomForest}
\title{Print method for random forest}
\usage{
\method{print}{spotRandomForest}(x, ...)
}
\arguments{
\item{...}{not used}

\item{object}{fit of the model, an object of class \code{"spotRandomForest"}, produced by \code{\link{buildRandomForest}}.}
}
\description{
Wrapper for \code{print.randomForest}.
}
\keyword{internal}
