% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRanger.R
\name{print.spotRanger}
\alias{print.spotRanger}
\title{Print method for random forest}
\usage{
\method{print}{spotRanger}(x, ...)
}
\arguments{
\item{...}{not used}

\item{object}{fit of the model, an object of class \code{"spotRandomForest"}, produced by \code{\link{buildRandomForest}}.}
}
\description{
Wrapper for \code{print.ranger}.
}
\keyword{internal}
