% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildRSM.R
\name{print.spotRSM}
\alias{print.spotRSM}
\title{Print method for RSM model}
\usage{
\method{print}{spotRSM}(x, ...)
}
\arguments{
\item{...}{not used}

\item{object}{fit of the model, an object of class \code{"spotRSM"}, produced by \code{\link{buildRSM}}.}
}
\description{
Wrapper for \code{summary.rsm}.
}
\seealso{
\code{\link{buildRSM}}
}
\keyword{internal}
