% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildSO.R
\name{print.spotSecondOrderModel}
\alias{print.spotSecondOrderModel}
\title{Print method for linear model}
\usage{
\method{print}{spotSecondOrderModel}(x, ...)
}
\arguments{
\item{...}{not used}

\item{object}{fit of the model, an object of class \code{"spotSecondOrderModel"}, produced by \code{\link{buildSO}}.}
}
\description{
Wrapper for \code{print.lm}.
}
\keyword{internal}
