% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSeverity.R
\name{print.spotSeverity}
\alias{print.spotSeverity}
\title{Print method for spotSeverity}
\usage{
\method{print}{spotSeverity}(x, ...)
}
\arguments{
\item{...}{not used}

\item{object}{an object of class \code{"spotSeverity"}, produced by \code{\link{spotSeverity}}.}
}
\description{
Wrapper for \code{print.spotSeverity}.
}
\keyword{internal}
