% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTree.R
\name{print.spotTreeModel}
\alias{print.spotTreeModel}
\title{Print method for rpart tree models}
\usage{
\method{print}{spotTreeModel}(x, ...)
}
\arguments{
\item{...}{not used}

\item{object}{fit of the model, an object of class \code{"spotRandomForest"}, produced by \code{\link{buildRandomForest}}.}
}
\description{
Wrapper for \code{print.rpart}.
}
\keyword{internal}
