% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildKrigingDACE.R
\name{regpoly1}
\alias{regpoly1}
\title{Regression: Regpoly1}
\usage{
regpoly1(S, grad = FALSE)
}
\arguments{
\item{S}{m*n matrix with \code{m} design points of dimension \code{n}}

\item{grad}{define if function returns gradient, default is \code{FALSE}}
}
\value{
returns a list with two elements:
			\item{\code{f}}{matrix of two columns: \cr 
		1. vector of ones with length \code{m} \cr
		2. \code{S}}
			\item{\code{df}}{Jacobian at the first point (first row in S) }
}
\description{
First order polynomial regression function.
}
\seealso{
\code{\link{buildKrigingDACE}}
}
\author{
The authors of the original DACE Matlab code
are Hans Bruun Nielsen, Soren Nymand Lophaven and Jacob Sondergaard. \cr
Ported to R by Martin Zaefferer \email{martin.zaefferer@fh-koeln.de}.
}
\keyword{internal}
