% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/repair.R
\name{repairNonNumeric}
\alias{repairNonNumeric}
\title{Repair Non-numeric Values}
\usage{
repairNonNumeric(x, types)
}
\arguments{
\item{x}{matrix to be rounded}

\item{types}{data types of the respective columns, numeric columns are specified by \code{"numeric"}.}
}
\description{
Round non-numeric columns of a matrix, specified by a vector of data given data types.
}
\examples{
x <- matrix(10*runif(12),4,3)
types <- c("numeric","factor","factor")
repairNonNumeric(x,types)
}
\keyword{internal}
