% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildKrigingDACE.R
\name{repmat}
\alias{repmat}
\title{repmat}
\usage{
repmat(a, n, m)
}
\arguments{
\item{a}{matrix}

\item{n}{dimension 1 (rows)}

\item{m}{dimension 2 (cols)}
}
\value{
Returns repeated matrix
}
\description{
Reproduce what MATLAB's repmat function can do, using kronecker function.
}
\keyword{internal}
