% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resBench01.R
\docType{data}
\name{resBench01}
\alias{resBench01}
\title{result from the vignette benchmark}
\format{
A list of 
\describe{
  \item{xbest}{num [1, 1:100] 188 45}
}
}
\usage{
resBench01
}
\description{
A data set 
The corresponding code can be found in the vignette SPOTVignetteNutshell.
}
\keyword{datasets}
