% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funSring.R
\name{ring}
\alias{ring}
\title{ring}
\usage{
ring(params)
}
\arguments{
\item{params}{list of 
\describe{
    \item{\code{randomSeed}}{random seed}
    \item{\code{nStates}}{number of S-Ring states}
    \item{\code{nElevators}}{number of elevators}
    \item{\code{probNewCustomer}}{probability pf a customer arrival}
    \item{\code{counter}}{Counter: number of waiting customers}
    \item{\code{sElevator}}{Vector representing elevators (s)}
     \item{\code{sCustomer}}{Vector representing customers (c)}
    \item{\code{currentState}}{Current state that is calculated}
    \item{\code{nextState}}{Next state that is calculated}
    \item{\code{nWeights}}{Number of weights for the perceptron (= 2 * nStates)}
    }}
}
\value{
number of waiting customers (estimation)
}
\description{
main function which iterates the ring
}
