% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/benchOptim.R
\name{runOptim}
\alias{runOptim}
\title{runOptim}
\usage{
runOptim(
  fl = makeMoreFunList(),
  method = "Nelder-Mead",
  n = 2,
  k = 1:length(makeMoreFunList()$funList),
  verbosity = 0
)
}
\arguments{
\item{fl}{function list. Generated with one of the function list 
generators in \code{spot}, e.g., \code{\link{makeSpotFunList}} or 
\code{\link{makeMoreFunList}}. Default: \code{\link{makeMoreFunList}}.}

\item{method}{The method used by \code{\link[stats]{optim}}: 
\code{"Nelder-Mead", "BFGS", "CG", "L-BFGS-B", "SANN", or "Brent"}.
Default: \code{"Nelder-Mead"}.}

\item{n}{repeats. If \code{n >1 }, different start points (randomized) 
will be used. Default: \code{n=2}.}

\item{k}{subset of benchmark functions.
Default: \code{1:length(makeMoreFunList()$funList)},
i.e., all implemented functions.}

\item{verbosity}{Level 0 shows no output (default).}
}
\value{
res. data.frame with results: \code{c("f", "r", "y")}
}
\description{
Run \code{\link[stats]{optim}} on a list of spot benchmark functions
}
\examples{
summary(runOptim(k=1)$y)
summary(runOptim(k=1, method="CG")$y)

}
