% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/shiny_app.R
\name{runSPOT}
\alias{runSPOT}
\title{Run the SPOT app}
\usage{
runSPOT(py_path = NULL)
}
\arguments{
\item{py_path}{path to the python executable. See decription for more information}
}
\description{
Runs the SPOT app. You will need python with libraries (numpy, osmnx, networkx, os, sklearn, pandas).
Pass the python executable path while calling the function.
To get the python path you can use

python3 -c "import sys; print(sys.executable)".

Or you can go to: https://stackoverflow.com/questions/749711/how-to-get-the-python-exe-location-programmatically
}
\examples{
runSPOT(py_path = "path/to/pyhton")
}
