% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectN.R
\name{selectN}
\alias{selectN}
\title{Select n Design Points}
\usage{
selectN(x, y, control)
}
\arguments{
\item{x}{matrix of design points}

\item{y}{matrix of function values}

\item{control}{list of controls
\describe{
    \item{\code{N}}{number of points to be returned.}
}}
}
\value{
This function returns a list with n design points:
\describe{
	\item{\code{x}}{matrix of selected design points.}
	\item{\code{y}}{ matrix of selected function values.}
}
}
\description{
This function returns a subset of n design points.
}
\keyword{internal}
