% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/package.R
\docType{package}
\name{SPOT-package}
\alias{SPOT-package}
\alias{SPOT}
\title{Sequential Parameter Optimization Toolbox}
\description{
Sequential Parameter Optimization Toolbox
}
\details{
SPOT uses a combination statistic models and optimization algorithms 
for the purpose of parameter optimization. Design of Experiment methods
are employed to generate an initial set of candidate solutions, which
are evaluated with a user-provided objective function.
The resulting data is used to fit a model, which in turn is subject
to an optimization algorithm, to find the most promising candidate solution(s).
These are again evaluated, after which the model is updated with the new results.
This sequential procedure of modeling, optimization, and evaluation is iterated
until the evaluation budget is exhausted.
}
\section{Maintainer}{

Thomas Bartz-Beielstein \email{tbb@bartzundbartz.de}
}

\seealso{
Main interface function is \code{\link{spot}}.
}
\author{
Thomas Bartz-Beielstein \email{tbb@bartzundbartz.de}, Martin Zaefferer, and F. Rehbach
with contributions from: C. Lasarczyk, M. Rebolledo, Joerg Stork.
}
\keyword{package}
