% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimES.R
\name{spotAlgEs}
\alias{spotAlgEs}
\title{Evolution Strategy Implementation}
\usage{
spotAlgEs(mue = 10, nu = 10, dimension = 2, mutation = 2,
  sigmaInit = 1, nSigma = 1, tau0 = 0, tau = 1, rho = "bi",
  sel = -1, stratReco = 1, objReco = 2, maxGen = Inf,
  maxIter = Inf, seed = 1, noise = 0, fName = funSphere,
  lowerLimit = -1, upperLimit = 1, verbosity = 0,
  plotResult = FALSE, logPlotResult = FALSE, sigmaRestart = 0.1,
  preScanMult = 1, globalOpt = NULL, ...)
}
\arguments{
\item{mue}{number of parents, default is \code{10}}

\item{nu}{selection pressure. That means, number of offspring (lambda) is mue multiplied with nu. Default is \code{10}}

\item{dimension}{dimension number of the target function, default is \code{2}}

\item{mutation}{mutation type, either \code{1} or \code{2}, default is \code{1}}

\item{sigmaInit}{initial sigma value (step size), default is \code{1.0}}

\item{nSigma}{number of different sigmas, default is \code{1}}

\item{tau0}{number, default is \code{0.0}. tau0 is the general multiplier.}

\item{tau}{number, learning parameter for self adaption, default is \code{1.0}. tau is the local multiplier for step sizes (for each dimension).}

\item{rho}{number of parents involved in the procreation of an offspring (mixing number), default is \code{"bi"}}

\item{sel}{number of selected individuals, default is \code{1}}

\item{stratReco}{Recombination operator for strategy variables. \code{1}: none. \code{2}: dominant/discrete (default). \code{3}: intermediate. \code{4}: variation of intermediate recombination.}

\item{objReco}{Recombination operator for object variables. \code{1}: none. \code{2}: dominant/discrete (default). \code{3}: intermediate. \code{4}: variation of intermediate recombination.}

\item{maxGen}{number of generations, stopping criterion, default is \code{Inf}}

\item{maxIter}{number of iterations (function evaluations), stopping criterion, default is \code{100}}

\item{seed}{number, random seed, default is \code{1}}

\item{noise}{number, value of noise added to fitness values, default is \code{0.0}}

\item{fName}{function, fitness function, default is \code{\link{funSphere}}}

\item{lowerLimit}{number, lower limit for search space, default is \code{-1.0}}

\item{upperLimit}{number, upper limit for search space, default is \code{1.0}}

\item{verbosity}{defines output verbosity of the ES, default is \code{0}}

\item{plotResult}{boolean, asks if results are plotted, default is \code{FALSE}}

\item{logPlotResult}{boolean, asks if plot results should be logarithmic, default is \code{FALSE}}

\item{sigmaRestart}{number, value of sigma on restart, default is \code{0.1}}

\item{preScanMult}{initial population size is multiplied by this number for a pre-scan, default is \code{1}}

\item{globalOpt}{termination criterion on reaching a desired optimum value, should be a vector of length dimension (LOCATION of the optimum). Default to NULL, which means it is ignored.}

\item{...}{additional parameters to be passed on to \code{fName}}
}
\description{
This function is used by \code{\link{optimES}} as a main loop for running
the Evolution Strategy with the given parameter set specified by SPOT.
}
