\name{spotAlgEsBanana}
\alias{spotAlgEsBanana}
\title{Rosenbrock Banana Function...}
\usage{spotAlgEsBanana(x, noise=0)}
\description{Rosenbrock Banana Function}
\details{This is a fitness function that could be used for the ES.}
\value{number \code{res} \cr
- \code{res} is the value of the corresponding \code{x} vector}
\arguments{\item{x}{vector to be evaluated by the fitness function}
\item{noise}{the noise is added to the result, defaults to zero}}
