\name{spotAlgEsBananaGradient}
\alias{spotAlgEsBananaGradient}
\title{Gradient Function of 'spotAlgEsBanana'...}
\usage{spotAlgEsBananaGradient(x)}
\description{Gradient Function of 'spotAlgEsBanana'}
\details{This is a fitness function that could be used for the ES.}
\value{number \code{res} \cr
- \code{res} is the value of the corresponding \code{x} vector}
\references{\code{\link{spotAlgEsBanana}}}
\arguments{\item{x}{vector to be evaluated by the fitness function}}
