\name{spotAlgEsInterReco}
\alias{spotAlgEsInterReco}
\title{spotAlgEsInterReco...}
\usage{spotAlgEsInterReco(parents, rhoVal, dimension, nSigma, objType="obj")}
\description{spotAlgEsInterReco}
\details{Recombination function for the ES}
\value{\code{rObject}}
\arguments{\item{parents}{Parent individuals}
\item{rhoVal}{number of parents involved in the procreation of an offspring}
\item{dimension}{number of dimensions}
\item{nSigma}{number of standard deviations}
\item{objType}{string, default is "obj"}}
