% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimES.R
\name{spotAlgEsInterRecoBeSw02}
\alias{spotAlgEsInterRecoBeSw02}
\title{spotAlgEsInterRecoBeSw02}
\usage{
spotAlgEsInterRecoBeSw02(parents, dimension, nSigma, objType = "obj")
}
\arguments{
\item{parents}{Parent individuals}

\item{dimension}{number of dimensions}

\item{nSigma}{number of standard deviations}

\item{objType}{string, default is "obj"}
}
\description{
Recombination function for the Evolution Strategy.
}
\seealso{
\code{\link{spotAlgEs}} \code{\link{spotAlgEsInterReco}} \code{\link{spotAlgEsDominantReco}}
}
\keyword{internal}
