\name{spotAlgEsMarriage}
\alias{spotAlgEsMarriage}
\title{Marriage...}
\usage{spotAlgEsMarriage(pop, rhoVal)}
\description{Marriage}
\details{Recombination function for the ES}
\value{\code{pop} \cr
- \code{pop} Population}
\arguments{\item{pop}{Population}
\item{rhoVal}{number of parents involved in the procreation of an offspring}}
