\name{spotAlgEsMarriageWithReplace}
\alias{spotAlgEsMarriageWithReplace}
\title{Marriage with replace...}
\usage{spotAlgEsMarriageWithReplace(pop, rhoVal)}
\description{Marriage with replace}
\details{Recombination function for the ES}
\value{\code{pop} \cr
- \code{pop} Population}
\arguments{\item{pop}{Population}
\item{rhoVal}{number of parents involved in the procreation of an offspring}}
