\name{spotAlgEsNoisyBraninFunction}
\alias{spotAlgEsNoisyBraninFunction}
\title{Noisy Branin Function...}
\usage{spotAlgEsNoisyBraninFunction(x, noise=0)}
\description{Noisy Branin Function}
\details{This is a fitness function that could be used for the ES.}
\value{number \code{res} \cr
- \code{res} is the value of the corresponding \code{x} vector}
\arguments{\item{x}{vector to be evaluated by the fitness function}
\item{noise}{the noise is added to the result, defaults to zero}}
