% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spotAlgStartEs.R
\name{spotAlgEsObjMutation}
\alias{spotAlgEsObjMutation}
\title{spotAlgEsObjMutation}
\usage{
spotAlgEsObjMutation(obj, strat)
}
\arguments{
\item{obj}{Object to be mutated}

\item{strat}{Strategy parameter sigma to mutate with}
}
\description{
Mutation function for the ES individual parameters
}
\keyword{internal}

