\name{spotAlgEsQuickTest}
\alias{spotAlgEsQuickTest}
\title{ES Quick Test}
\usage{
  spotAlgEsQuickTest(mue = 2, nu = 2, tau = 1)
}
\arguments{
  \item{mue}{value of mue (number of parents) to be used
  with the ES default is 2}

  \item{nu}{value of nu to be used with the ES default is
  2}

  \item{tau}{value of tau (learning parameter) to be used
  with the ES default is 1}
}
\description{
  This is a quick test function for the Evolution Strategy
  algorithm.
}
\seealso{
  \code{\link{spotAlgEs}} \code{\link{spotAlgStartEs}}
}
\keyword{internal}

