\name{spotAlgEsRastrigin}
\alias{spotAlgEsRastrigin}
\title{rastrigin...}
\usage{spotAlgEsRastrigin(x)}
\description{rastrigin}
\details{This is a fitness function that could be used for the ES.}
\value{number \code{res} \cr
- \code{res} is the value of the corresponding \code{x} vector}
\arguments{\item{x}{vector to be evaluated by the fitness function}}
