\name{spotAlgEsSelection}
\alias{spotAlgEsSelection}
\title{spotAlgEsSelection...}
\usage{spotAlgEsSelection(parentPop, offspringPop, sel, gen, iter, maxIter)}
\description{spotAlgEsSelection}
\details{Selection function for the ES}
\arguments{\item{parentPop}{Parent population}
\item{offspringPop}{Offspring population}
\item{sel}{number of individuals to be selected}
\item{gen}{generation number}
\item{iter}{current iteration number}
\item{maxIter}{maximum iteration number}}
