\name{spotAlgEsSphere}
\alias{spotAlgEsSphere}
\title{Sphere Function...}
\usage{spotAlgEsSphere(x)}
\description{Sphere Function}
\details{This is a fitness function that could be used for the ES which is an example algorithm to be optimized by SPOT.}
\value{number \code{res} \cr
- \code{res} is the function value of the corresponding vector \code{x}}
\references{\code{\link{spotAlgEsSphere1}} \code{\link{spotAlgEsF}}}
\arguments{\item{x}{vector to be evaluated by the fitness function}}
