\name{spotAlgEsSphere1}
\alias{spotAlgEsSphere1}
\title{Sphere Function...}
\usage{spotAlgEsSphere1(x)}
\description{Sphere Function}
\details{This is a fitness function that could be used for the ES.}
\value{number \code{res} \cr
- \code{res} is the value of the corresponding \code{x} vector}
\references{\code{\link{spotAlgEsSphere}}}
\arguments{\item{x}{vector to be evaluated by the fitness function}}
