\name{spotAlgEsWild}
\alias{spotAlgEsWild}
\title{Wild Function...}
\usage{spotAlgEsWild(x, noise=0)}
\description{Wild Function}
\details{This is a fitness function that could be used for the ES which is an example algorithm to be optimized by SPOT.}
\value{number \code{res} \cr
- \code{res} is the function value of the corresponding vector \code{x}}
\references{\code{\link{spotAlgEsF}}}
\arguments{\item{x}{vector to be evaluated by the fitness function}
\item{noise}{the noise is added to the result}}
