% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotCreateDesignFactors.R
\name{spotCreateDesignFactors}
\alias{spotCreateDesignFactors}
\title{spotCreateDesignFactors}
\usage{
spotCreateDesignFactors(spotConfig, noDesPoints = NaN, repeats = NaN)
}
\arguments{
\item{spotConfig}{list of spot settings}

\item{noDesPoints}{is obsolete for this type of design, it has no influence. The number of points is fixed by the number of factors and their levels.}

\item{repeats}{is obsolete for this type of design, it has no influence.}
}
\value{
matrix \code{M} \cr
- \code{M} has \code{dimension} columns and \code{noDesPoints} rows
}
\description{
Design function for designs consisting of factors only. Factorial designs for al factors.
Assumse that factors are encoded in the ROI, with integer notation.
}
\seealso{
\code{\link{spotCreateDesignBasicDoe}}, \code{\link{spotCreateDesignLhd}}, 
\code{\link{spotCreateDesignLhs}}, \code{\link{spotCreateDesignLhsOpt}}
}

