% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotCreateDesignLhsOpt.R
\name{spotCreateDesignLhsOpt}
\alias{spotCreateDesignLhsOpt}
\title{spotCreateDesignLhsOpt}
\usage{
spotCreateDesignLhsOpt(spotConfig, noDesPoints = NaN, repeats = NaN)
}
\arguments{
\item{spotConfig}{list of spotConfiguration}

\item{noDesPoints}{number of design points, default is NaN}

\item{repeats}{number of repeats, default is NaN}
}
\value{
matrix \code{M} \cr
- \code{M} has \code{dimension} columns and \code{noDesPoints} rows
with entries corresponding to the region of interest.
}
\description{
Uses the optimumLHS function from the lhs package to create a Latin Hypercube
Design. optimumLHS uses the S optimality criterion.
}
\seealso{
\code{\link{spotCreateDesignBasicDoe}}, \code{\link{spotCreateDesignFrF2}}, 
\code{\link{spotCreateDesignLhd}}, \code{\link{spotCreateDesignLhsOpt}}
}

