\name{spotEnsembleMinimum}
\alias{spotEnsembleMinimum}
\title{spotEnsembleMinimum}
\usage{
  spotEnsembleMinimum(x)
}
\arguments{
  \item{x}{result structure: a list with element \code{Y}
  for predicted values and \code{fits} for model fit list}
}
\value{
  the summed
}
\description{
  Minn over all points.
}
\keyword{internal}

