% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotEnsembleFunctions.R
\name{spotEnsembleModelBuilding}
\alias{spotEnsembleModelBuilding}
\title{spotEnsembleModelBuilding}
\usage{
spotEnsembleModelBuilding(rawB, mergedB, design, spotConfig)
}
\arguments{
\item{rawB}{unmerged data, passed to the model function (i.e. \code{spotPredict*})}

\item{mergedB}{merged data, passed to the model function (i.e. \code{spotPredict*}}

\item{design}{new design points which should be predicted}

\item{spotConfig}{global list of all options}
}
\value{
returns a list, containing the element \code{Y} which is a list of results from each mode, and the element \code{fits} which is a list of model fits.
}
\description{
This function is used in the multi ensembles (e.g. \code{\link{spotEnsembleMultiAlternate}}) to
build and evaluate the models in the ensemble.
}
\keyword{internal}

