\name{spotEnsembleStatsMin}
\alias{spotEnsembleStatsMin}
\title{spotEnsembleStatsMin}
\usage{
  spotEnsembleStatsMin(x)
}
\arguments{
  \item{x}{result structure: a list with element \code{Y}
  for predicted values and \code{fits} for model fit list}
}
\value{
  returns a vector of results, with one element of the
  vector for each fit
}
\description{
  Get minimum of ensemble results.
}
\keyword{internal}

