% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spotExpectedImprovement.R
\name{spotExiPsi}
\alias{spotExiPsi}
\title{spotExiPsi}
\usage{
spotExiPsi(a, b, m, s)
}
\arguments{
\item{a}{a}

\item{b}{b}

\item{m}{m}

\item{s}{s}
}
\description{
helper function for spotSExI2d
}
\keyword{internal}

