% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot.R
\name{spotFillControlList}
\alias{spotFillControlList}
\title{spotFillControlList}
\usage{
spotFillControlList(controlList, lower, upper)
}
\arguments{
\item{controlList}{list of controls, see \code{\link{spotControl}}.}
}
\value{
a list
}
\description{
Fill in some values for the control list. Internal use only.
}
\keyword{internal}
