\name{spotFuncStartBranin}
\alias{spotFuncStartBranin}
\title{Branin function call for SPOT...}
\usage{spotFuncStartBranin(pdFile, desFileName, resFileName)}
\description{Branin function call for SPOT}
\details{SPOT uses this function for some demos to call the \code{\link{spotNoisyBraninFunction}} function}
\references{\code{\link{SPOT}} \code{\link{spot}} \code{\link{demo}} \code{\link{optim}}
\code{\link{spotNoisyBraninFunction}}}
\arguments{\item{pdFile}{name of the apd file}
\item{desFileName}{name of the des file}
\item{resFileName}{name of the res file}}
