\name{spotFuncStartBraninSann}
\alias{spotFuncStartBraninSann}
\title{Branin function...}
\usage{spotFuncStartBraninSann(x)}
\description{Branin function}
\details{This is the implementation of the Branin function used by some SPOT demos

This Branin function does not use noise}
\value{number \code{y} \cr
- \code{y} is the value of the corresponding \code{x} vector}
\references{\code{\link{SPOT}} \code{\link{spot}} \code{\link{demo}} \code{\link{spotNoisyBraninFunction}}
\code{\link{spotFuncStartBranin}}}
\arguments{\item{x}{two dimensional vector that will be evaluated by the Branin function}}
