\name{spotFuncStartSixHump}
\alias{spotFuncStartSixHump}
\title{Six Hump function call for SPOT...}
\usage{spotFuncStartSixHump(pdFile, desFileName, resFileName)}
\description{Six Hump function call for SPOT}
\details{SPOT uses this function for some demos to call the \code{\link{spotNoisySixHumpFunction}} function}
\references{\code{\link{SPOT}} \code{\link{spot}} \code{\link{demo}} \code{\link{spotFuncStartBranin}}
\code{\link{spotFuncStartBranin}} \code{\link{spotNoisySixHumpFunction}}}
\arguments{\item{pdFile}{name of the apd file}
\item{desFileName}{name of the des file}
\item{resFileName}{name of the res file}}
