\name{spotFuncStartSphere}
\alias{spotFuncStartSphere}
\title{Sphere function call for SPOT...}
\usage{spotFuncStartSphere(pdFile, desFileName, resFileName)}
\description{Sphere function call for SPOT}
\details{SPOT uses this function for some demos to call the \code{\link{spotNoisySphereFunction}} function}
\references{\code{\link{SPOT}} \code{\link{spot}} \code{\link{demo}} \code{\link{optim}}
\code{\link{spotNoisySphereFunction}}}
\arguments{\item{pdFile}{name of the apd file}
\item{desFileName}{name of the des file}
\item{resFileName}{name of the res file}}
