\name{spotGetAllRepeats}
\alias{spotGetAllRepeats}
\title{Get ALL Repeats Function}
\usage{
  spotGetAllRepeats(configList)
}
\arguments{
  \item{configList}{config list}
}
\value{
  list \code{all} \cr - \code{all} contains all repeats of
  the given config list
}
\description{
  Gets repeats of all configs in the list
}
\keyword{internal}

