% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spotGui.R
\name{spotGui}
\alias{spotGui}
\title{Start the SPOT GUI}
\usage{
spotGui()
}
\description{
This function starts the graphical user interface for SPOT, which
is based on java. The .jar File started with this function is
in the directory where SPOT is installed.
}
\details{
Java runtime environment needs to be installed to use the GUI.
}

