% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spotPredictDace.R
\name{spotHelpBslash}
\alias{spotHelpBslash}
\title{Backslash operator.}
\usage{
spotHelpBslash(X, Y)
}
\arguments{
\item{X}{X matrix}

\item{Y}{Y vector}
}
\value{
Returns coefficients
}
\description{
Reproduce what MATLAB's backslash operator can do, using qr() and qr.coef().
}
\keyword{internal}

