\name{spotHelpInterfaces...}
\alias{spotHelpInterfaces...}
\title{General Help on SPOT...}
\usage{
  spotHelpInterfaces...()
}
\description{
  SPOT is a system open and designed to plug in \cr - your
  own algorithm see \code{\link{spotAlgStart...}}\cr - your
  own predictors see \code{\link{spotPredict...}}\cr - your
  own design creating functions
  \code{\link{spotCreateDesign...}}\cr - your own increase
  functions to increase the repeats for each sequential
  step
}

