% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spotStatistics.R
\name{spotHlpF.norminv}
\alias{spotHlpF.norminv}
\title{Helper function for transformations/normalizing}
\usage{
spotHlpF.norminv(x, z)
}
\arguments{
\item{x}{parameter vector}

\item{z}{vector (zmin, zmax)}
}
\value{
vector \code{y} \cr
- \code{y} is the transformed/normalized version of \code{x}
}
\description{
Transforms natural to coded variables
}
\keyword{internal}

