% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spotExpectedImprovement.R
\name{spotInfillProbImp}
\alias{spotInfillProbImp}
\title{Probability of Improvement Infill Criterion}
\usage{
spotInfillProbImp(mean, sd, min)
}
\arguments{
\item{mean}{predicted mean values}

\item{sd}{predicted st. deviation}

\item{min}{the currently known optimum value}
}
\value{
Returns the negative of the Probability of Improvement.
}
\description{
This is the single objective infill criterion "Probability of Improvement" as introduced by Forrester (2008).
}
\references{
Forrester, Alexander I.J.; Sobester, Andras; Keane, Andy J. (2008). Engineering Design via Surrogate Modelling - A Practical Guide. John Wiley & Sons.
}

