\name{spotInterface...}
\alias{spotInterface...}
\title{SPOT Interface.}
\description{SPOT Interface...}
\details{SPOT is a system open and designed to plug in \cr 
- your own algorithm see  \code{\link{spotAlgStartAlg...}}\cr
- your own predictors see  \code{\link{spotPredict...}}\cr
- your own design creating functions  \code{\link{spotCreate...}}\cr
- your own increase functions to increase the repeats for each sequential step}
