% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotEnsembleSingleBLAnorm.R
\name{spotInvChisq}
\alias{spotInvChisq}
\title{Inverse Chisquared distribution}
\usage{
spotInvChisq(n, df, scale)
}
\arguments{
\item{n}{number of observations}

\item{df}{degrees of freedom}

\item{scale}{scale}
}
\value{
Inverse Chisquare
}
\description{
Inverse Chisquared distribution
}
\keyword{internal}

