% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spotMcoSort.R
\name{spotMcoCrowdTournament}
\alias{spotMcoCrowdTournament}
\title{Multi criteria sorting by a tournament scheme}
\usage{
spotMcoCrowdTournament(largeDesign, designY, tsize, ssize)
}
\arguments{
\item{largeDesign}{the design matrix in the parameter space, to be sorted by the associated y-values for each objective}

\item{designY}{objective value matrix. Contains objective values associated to largeDesign}

\item{tsize}{tournament size (percentage of large design in values between zero and one)}

\item{ssize}{selection size number of points that will be selected in SPOT, lower bound for absolute tsize}
}
\value{
largeDesign \cr
- The sorted large design
}
\description{
Experimental, do not use
}
\keyword{internal}

