\name{spotMetaGetAssOp}
\alias{spotMetaGetAssOp}
\title{Spot Meta Get Assignment Operator...}
\usage{spotMetaGetAssOp(metaFile)}
\description{Spot Meta Get Assignment Operator}
\details{Help function that simply searches for the assignment operator in the 
.meta-file and returns the value or stops}
\value{character \cr
- holding either "=" or "<-" respectively}
\arguments{\item{metaFile}{file that holds the meta - variables that should be 
added full factorial to the apd-files}}
