\name{spotMetaGetMetaVarNames}
\alias{spotMetaGetMetaVarNames}
\title{Spot Meta Get Meta Var Names...}
\usage{spotMetaGetMetaVarNames(spotConfig)}
\description{Spot Meta Get Meta Var Names}
\details{Returns the names of the meta variables as they will appear in the FBS file.
Note that an N-dimensional vector x will appear as x0, x1, ..., xN in the FBS
file.}
\value{character vector \cr
- The names of the meta variables as they will appear in the FBS file.}
\arguments{\item{spotConfig}{spotConfig$io.metaFileName is used.}}
