\name{spotMetaRead}
\alias{spotMetaRead}
\title{Spot Meta Read...}
\usage{spotMetaRead(metaFile)}
\description{Spot Meta Read}
\details{Help function that simply reads the .metafile and returns the values in a list}
\value{\code{myList} \cr
- \code{myList} is holding a list of all variables defined in the .meta file}
\arguments{\item{metaFile}{file that holds the meta - variables that should be 
added full factorial to the apd-files}}
