\name{spotMexicanHatFunction}
\alias{spotMexicanHatFunction}
\title{noisy Mexican Hat function...}
\usage{spotMexicanHatFunction(x)}
\description{noisy Mexican Hat function}
\details{This is the noisy implementation of the Mexican Hat function used by some SPOT demos}
\value{number \code{y} \cr
- \code{y} is the function value of the corresponding vector \code{x}}
\references{\code{\link{SPOT}} \code{\link{spot}} \code{\link{demo}} \code{\link{spotFuncStartMexicanHat}}
\code{\link{spotFuncStartMexicanHat}}  \code{\link{rnorm}} \code{\link{spotAlgEsF}}}
\arguments{\item{x}{two dimensional vector that will be evauluated by the branin function}}
