\name{spotNoisyBraninFunction}
\alias{spotNoisyBraninFunction}
\title{noisy Branin function...}
\usage{spotNoisyBraninFunction(x, noise)}
\description{noisy Branin function}
\details{This is the noisy implementation of the Branin function used by some SPOT demos}
\value{number \code{y} \cr
- \code{y} is the value of the corresponding \code{x} vector}
\references{\code{\link{SPOT}} \code{\link{spot}} \code{\link{demo}} \code{\link{spotFuncStartBranin}}
\code{\link{spotFuncStartBranin}}}
\arguments{\item{x}{two dimensional vector that will be evauluated by the branin function}
\item{noise}{this number is added to the function value y as noise}}
