\name{spotNoisySixHumpFunction}
\alias{spotNoisySixHumpFunction}
\title{noisy Six Hump function...}
\usage{spotNoisySixHumpFunction(x, noise=0)}
\description{noisy Six Hump function}
\details{Example for SPOT optimizing the 6-hump camel back function, see 
http://www.it.lut.fi/ip/evo/functions/node26.html}
\value{number \code{y} \cr
- \code{y} is the value of the corresponding \code{x} vector}
\references{\code{\link{SPOT}} \code{\link{spot}} \code{\link{demo}} \code{\link{spotFuncStartBranin}}
\code{\link{spotFuncStartBranin}} \code{\link{spotFuncStartSixHump}}}
\arguments{\item{x}{two dimensional vector that will be evauluated by the branin function}
\item{noise}{this number is added to the function value y as noise}}
