\name{spotNoisySphereFunction}
\alias{spotNoisySphereFunction}
\title{3-dim sphere function...}
\usage{spotNoisySphereFunction(x, noise)}
\description{3-dim sphere function}
\details{This is the noisy implementation of the Sphere function used by some SPOT demos}
\value{number \code{y} \cr
- \code{y} is the value of the corresponding \code{x} vector}
\references{\code{\link{SPOT}} \code{\link{spot}} \code{\link{demo}} \code{\link{spotFuncStartSphere}}
\code{\link{spotFuncStartSphere}}}
\arguments{\item{x}{two dimensional vector that will be evauluated by the sphere function}
\item{noise}{this number is added to the function value y as noise}}
