% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spotModelParetoOptim.R
\name{spotParetoOptMulti}
\alias{spotParetoOptMulti}
\title{DEPRECATED}
\usage{
spotParetoOptMulti(startPoint, spotConfig)
}
\arguments{
\item{startPoint}{see \code{\link{spotModelParetoOptim}}}

\item{spotConfig}{see \code{\link{spotModelParetoOptim}}}
}
\description{
see \code{\link{spotModelParetoOptim}}
}
\keyword{internal}

