% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotRepairMissingValues.R
\name{spotPenalizeMissingValues}
\alias{spotPenalizeMissingValues}
\title{Penalize Missing Values}
\usage{
spotPenalizeMissingValues(x, y, nafit, penalty)
}
\arguments{
\item{x}{data frame of locations}

\item{y}{vector of observations}

\item{nafit}{randomForest fit (object), that models dependency of NAs on x}

\item{penalty}{(usually positive) penalty value}
}
\value{
\code{y}: Observations + penalty where applicable
}
\description{
Values that are estimated to yield NAs will be penalized.
}
\seealso{
\code{\link{spotRepairMissingValues}}
}

