\name{spotPlotBst}
\alias{spotPlotBst}
\title{Plot Best...}
\usage{spotPlotBst(spotConfig)}
\description{Plot Best}
\details{Function  used to continuosly plot the actually retrieved best value 
throughout a SPOT run. The number of variables shown is limited to twelve -
make sure the relevant variables belong to the first lines of your .roi-file.}
\arguments{\item{spotConfig}{the list of all parameters is given, used here 
\code{alg.roi}: the region of interest reduced to a matrix}}
