\name{spotPlotPower}
\alias{spotPlotPower}
\title{spotPlotPower}
\usage{
  spotPlotPower(y0, y1, alpha = 0.05, add = FALSE, n = NA,
    rightLimit = 1)
}
\arguments{
  \item{y0}{First input vector}

  \item{y1}{Second input vector}

  \item{alpha}{description of alpha, default value is
  \code{0.05}}

  \item{add}{Boolean, default value is \code{FALSE}}

  \item{n}{number of vector elements that should be
  evaluated, default value is \code{NA}, which means the
  whole vector}

  \item{rightLimit}{description of rightLimit, default
  value is \code{1}}
}
\value{
  description of return value
}
\description{
  short description
}
\details{
  details
}
\seealso{
  \code{\link{spotPower}} \code{\link{spotCompare}}
}
\keyword{internal}

