\name{spotPlotSeverity}
\alias{spotPlotSeverity}
\title{spotPlotSeverity}
\usage{
  spotPlotSeverity(y0, y1, add = FALSE, n = NA, alpha,
    rightLimit = 1)
}
\arguments{
  \item{y0}{first input vector}

  \item{y1}{second input vector}

  \item{add}{default value is FALSE}

  \item{n}{default value is NA, which means length of y0
  will be used for n}

  \item{alpha}{description}

  \item{rightLimit}{description of rightLimit, default
  value is \code{1}}
}
\value{
  description of return value
}
\description{
  short description
}
\details{
  details
}
\seealso{
  \code{\link{spotSeverity}} \code{\link{spotCompare}}
}
\keyword{internal}

