% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSeverity.R
\name{spotPlotSeverityBasic}
\alias{spotPlotSeverityBasic}
\title{spotPlotSeverityBasic}
\usage{
spotPlotSeverityBasic(y0, y1, add = FALSE, n = NA, alpha, rightLimit = 1)
}
\arguments{
\item{y0}{first input vector}

\item{y1}{second input vector}

\item{add}{default value is FALSE}

\item{n}{default value is NA, which means length of y0 will be used for n}

\item{alpha}{description}

\item{rightLimit}{description of rightLimit, default value is \code{1}}
}
\value{
description of return value
}
\description{
spotPlotSeverityBasic
}
\examples{
### Example from D G Mayo and A Spanos. 
### Severe Testing as a Basic Concept in a Neyman–Pearson Philosophy of Induction. 
### British Journal for the Philosophy of Science, 57:323–357, 2006. (fig 2):
x0 <- 12.1
mu1 <- seq(11.9,13,0.01)
n <- 100
sigma <- 2
alpha <- 0.025
plot(mu1, spotSeverityBasic(x0, mu1, n, sigma, alpha), type = "l", ylim=c(0,1), col="blue")
abline(h=0)
abline(h=1)
 abline(h=0.95)
abline(v=12.43)
### plot power:
mu0 <- 12
points(mu1, spotPower(alpha, mu0, mu1, n, sigma), type = "l", ylim=c(0,1), col="green")
abline(v=12.72)

}
