\name{spotPower}
\alias{spotPower}
\title{spotPower...}
\usage{spotPower(alpha, mu0, mu1, n, sigma)}
\description{spotPower}
\details{short description

details}
\value{description of return value}
\references{\code{\link{spotPlotPower}} 
\code{\link{spotCompare}}}
\arguments{\item{alpha}{description of alpha}
\item{mu0}{description of mu0}
\item{mu1}{description of mu1}
\item{n}{vector length}
\item{sigma}{standart deviation}}
