% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotPredictEsvm.R
\name{spotPredictEsvm}
\alias{spotPredictEsvm}
\title{Meta Model Interface: Support Vector Machine}
\usage{
spotPredictEsvm(rawB, mergedB, design, spotConfig, fit = NULL)
}
\arguments{
\item{rawB}{unmerged data}

\item{mergedB}{merged data}

\item{design}{new design points which should be predicted}

\item{spotConfig}{global list of all options, needed to provide data for calling functions}

\item{fit}{if an existing model fit is supplied, the model will not be build based on 
data, but only evaluated with the model fit (on the design data). To build the model, 
this parameter has to be NULL. If it is not NULL the parameters mergedB and rawB will not be 
used at all in the function.}
}
\value{
returns the list \code{spotConfig} with two new entries:\cr
	spotConfig$seq.modelFit fit of the model used with predict() \cr
spotConfig$seq.largeDesignY the y values of the design, evaluated with the fit
}
\description{
Meta model based on svm function in the e1071-package, which builds a
support vector machine for regression.
}

