\name{spotPredictOptMulti}
\alias{spotPredictOptMulti}
\title{DEPRECATED}
\usage{
  spotPredictOptMulti(startPoint, spotConfig)
}
\arguments{
  \item{startPoint}{see \code{\link{spotModelOptim}}}

  \item{spotConfig}{see \code{\link{spotModelOptim}}}
}
\description{
  see \code{\link{spotModelOptim}}
}
\keyword{internal}

