% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spotFuncTools.R
\name{spotPrepareResult}
\alias{spotPrepareResult}
\title{Write Result File}
\usage{
spotPrepareResult(y, colNames, res)
}
\arguments{
\item{y}{result value passed from target function}

\item{colNames}{column names of the results in the target data frame}

\item{res}{target data frame}
}
\value{
list \cr
containing the extended res data frame and the corrected colNames
}
\description{
Writes a new row to the result file and adds it to the existing data.frame
}
\keyword{internal}

