\name{spotPrint}
\alias{spotPrint}
\title{Spot Print...}
\usage{spotPrint(set.io.verbosity, io.verbosity, myArg)}
\description{Spot Print}
\details{This help function prints the given argument "myArg" 
only if user gives the io.verbosity to do so}
\references{\code{\link{SPOT}} \code{\link{writeLines}}}
\arguments{\item{set.io.verbosity}{\code{spotConfig$io.verbosity} should be passed here. Global flag to drive the \code{io.verbosity} of the programm}
\item{io.verbosity}{\code{io.verbosity} for this specified output string}
\item{myArg}{the argument to be printed}}
