% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spotRepairMissingValues.R
\name{spotRepairMissingValuesCoKriging}
\alias{spotRepairMissingValuesCoKriging}
\title{Repair Missing Values}
\usage{
spotRepairMissingValuesCoKriging(xe, ye, xc, yc, spotConfig, fitC)
}
\arguments{
\item{xe}{data frame of locations (expensive, fine function)}

\item{ye}{vector of observations (expensive, fine function)}

\item{xc}{data frame of locations (cheap, coarse function)}

\item{yc}{vector of observations (cheap, coarse function)}

\item{spotConfig}{configuration, including spotConfig$seq.na.penalty, spotConfig$seq.predictionModel.func and spotConfig$seq.modelFit}

\item{fitC}{Fit of a Kriging model (from \code{\link{forrBuilder}})}
}
\value{
\code{y}: Observations + penalty where applicable
}
\description{
Similar to \code{\link{spotRepairMissingValues}}, but concerns the case of
multi-fidelity optimization with Co-Kriging only.
}
\seealso{
\code{\link{spotRepairMissingValues}}
}

