% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotReportContour.R
\name{spotReportContourMulti}
\alias{spotReportContourMulti}
\title{Multi Objective Contour Plot}
\usage{
spotReportContourMulti(spotConfig)
}
\arguments{
\item{spotConfig}{the configuration list of all spot parameters. \cr
The parameter spotConfig$report.interactive=TRUE will be set as default if not contained in the list.
That means, by default the user will be asked to specify which parameters will be varied when the report is started. This is done in a small twiddler GUI. 
If the user wants to specify which parameters should be plotted against each other, before starting the report, he can 
set the parameter spotConfig$report.aIndex and spotConfig$report.bIndex. They should be two different integer numbers. They will only be used if 
spotConfig$report.interactive is FALSE. By default they will be set to 1 and 2, so the first two parameters in the ROI will be plotted.}
}
\description{
This is the multi objective version of spotReportContour. It is called by spotReportContour if more than one objective are optimized by SPOT. No need to call this as a user.
}
\keyword{internal}

