\name{spotReportSens}
\alias{spotReportSens}
\title{Sensitivity Report Function...}
\usage{spotReportSens(spotConfig)}
\description{Sensitivity Report Function}
\details{Function to generate a report with sensitivity plot.

The sensitivity curves are based on a metamodel which is a random forest with 100 trees 
fitted to the results}
\references{\code{\link{SPOT}} \code{\link{spot}} \code{\link{spotStepReport}}}
\arguments{\item{spotConfig}{the configuration list of all spot parameters}}
