\name{spotSeqDesignIncreaseMultTwo}
\alias{spotSeqDesignIncreaseMultTwo}
\title{Increase Sequential Design Mult Two...}
\usage{spotSeqDesignIncreaseMultTwo(actRepeats)}
\description{Increase Sequential Design Mult Two}
\details{This function increases the repeats done in each sequential designby multiplying by 2}
\value{number \code{repeats} \cr
- \code{repeats} is the new recommended number of repeats}
\references{\code{\link{SPOT}} \code{\link{spot}} \code{\link{spotSeqDesignIncreasePlusOne}}}
\arguments{\item{actRepeats}{must hold the number of repeats recommended last time.}}
