\name{spotSeverity}
\alias{spotSeverity}
\title{spotSeverity...}
\usage{spotSeverity(x0, mu1, n, sigma, alpha)}
\description{spotSeverity}
\details{short description

details}
\value{description of return value}
\references{\code{\link{spotPlotSeverity}}  \code{\link{spotCompare}}}
\arguments{\item{x0}{sample mean value}
\item{mu1}{description}
\item{n}{description}
\item{sigma}{description}
\item{alpha}{description}}
