\name{spotSixHumpFunction}
\alias{spotSixHumpFunction}
\title{noisy Six Hump function...}
\usage{spotSixHumpFunction(x)}
\description{noisy Six Hump function}
\details{Example for SPOT optimizing the 6-hump camel back function, see 
http://www.it.lut.fi/ip/evo/functions/node26.html}
\value{number \code{y} \cr
- \code{y} is the function value of the corresponding \code{x} vector. Min function value: -1.031628453}
\references{\code{\link{SPOT}} \code{\link{spot}} \code{\link{demo}} \code{\link{spotFuncStartBranin}}
\code{\link{spotFuncStartBranin}} \code{\link{spotFuncStartSixHump}}}
\arguments{\item{x}{two dimensional vector that will be evauluated by the branin function}}
