\name{spotSphereFunction}
\alias{spotSphereFunction}
\title{multi-dim sphere function...}
\usage{spotSphereFunction(x)}
\description{multi-dim sphere function}
\details{This is the noisy implementation of the Sphere function used by some SPOT demos}
\value{number \code{y} \cr
- \code{y} is the value of the corresponding \code{x} vector}
\references{\code{\link{SPOT}} \code{\link{spot}} \code{\link{demo}} \code{\link{spotFuncStartSphere}}
\code{\link{spotFuncStartSphere}}}
\arguments{\item{x}{two dimensional vector that will be evauluated by the branin function}}
