% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotHelpFunctions.R
\name{spotVersion}
\alias{spotVersion}
\title{Spot Version}
\usage{
spotVersion()
}
\value{
string \cr
holding the installed version of SPOT
}
\description{
Help function that returns the version of SPOT 
provided for convenience
}
\keyword{internal}

