% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spotHelpFunctions.R
\name{spotWriteAroi}
\alias{spotWriteAroi}
\title{Spot Write Aroi}
\usage{
spotWriteAroi(aroi, verbosity, sep, filename)
}
\arguments{
\item{aroi}{data frame.}

\item{verbosity}{for values greater than two, a message is given}

\item{sep}{the column separator used when writing the table to .aroi-file}

\item{filename}{the filename the aroi should be written to}
}
\description{
help function spotWriteAroi writes actual region of interest to the .aroi-file
}
\details{
Result/Effects:
rewrites the actual region of interest-file
}

