% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotHelpFunctions.R
\name{spotWriteLines}
\alias{spotWriteLines}
\title{Spot Write Lines}
\usage{
spotWriteLines(set.io.verbosity, io.verbosity, myString, con = stderr())
}
\arguments{
\item{set.io.verbosity}{\code{spotConfig$io.verbosity} should be passed here. Global flag to drive the \code{io.verbosity} of the program}

\item{io.verbosity}{\code{io.verbosity} for this specified output string}

\item{myString}{the string to be written to stdout}

\item{con}{defines the output stream, defaults to \code{stderr()}}
}
\description{
This help function writes the string given in "myString" 
only if user gives the io.verbosity to do so
}
\seealso{
\code{\link{SPOT}} \code{\link{writeLines}}
}
\keyword{internal}

