% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotFuncTools.R
\name{spotWriteRes}
\alias{spotWriteRes}
\title{Write Result File}
\usage{
spotWriteRes(res, file, mode, cres)
}
\arguments{
\item{res}{result to write}

\item{file}{name of result file}

\item{mode}{only write to file if mode is TRUE}

\item{cres}{res will be appended to this existing data frame}
}
\value{
numeric \cr
holding the noise Value
}
\description{
Writes a new row to the result file and adds it to the existing data.frame
}
\keyword{internal}

