% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spotHelpFunctions.R
\name{spotWriteRoiFileForRepeats}
\alias{spotWriteRoiFileForRepeats}
\title{write ROI for repeats}
\usage{
spotWriteRoiFileForRepeats(workingDir, bstFile, roiInFile, roiOutFile)
}
\arguments{
\item{workingDir}{Working directory as string}

\item{bstFile}{best file}

\item{roiInFile}{ROI in file}

\item{roiOutFile}{ROI out file}
}
\description{
write ROI for repeats
}
\keyword{internal}

