% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sringRes1.R
\docType{data}
\name{sringRes1}
\alias{sringRes1}
\title{S-Ring Simulation Data}
\format{
A data frame with 20 obs. of  3 variables:
\describe{
  \item{y}{num  10 10 10 10 10 ...}
  \item{sigma}{num  0.1 0.1 0.1 0.1 0.1 1 1 1 1 1 ..}
  \item{ne}{num  5 5 5 5 5 5 5 5 5 5 ...}
}
}
\usage{
sringRes1
}
\description{
A data set based on evaluations of the \code{funCosts} function.
The corresponding code can be found in the vignette SPOTVignetteElevator
}
\keyword{datasets}
