% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildBO.R
\name{thetaNugget}
\alias{thetaNugget}
\title{thetaNugget}
\usage{
thetaNugget(par, X, Y)
}
\arguments{
\item{par}{parameter vector. First dim(x) entries are theta values, 
last entry is nugget parameter.}

\item{X}{x coordinates}

\item{Y}{y values at x}
}
\value{
negLogLikelihood
}
\description{
get theta (distance, lengthscale)
and nugget (noise) parameters gradient
}
