% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotTools.R
\name{vmessage}
\alias{vmessage}
\title{formatted output dependent on verbosity}
\usage{
vmessage(verbosity, text, value)
}
\arguments{
\item{verbosity}{verbosity level}

\item{text}{output to be printed}

\item{value}{value to be printed}
}
\description{
Combine \code{\link{sprintf}} and \code{\link{writeLines}} to
generate formatted output
}
\examples{

x <- 123
vmessage(1, "value of x:" , x)

}
