% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{wrapBatchTools}
\alias{wrapBatchTools}
\title{wrapBatchTools}
\usage{
wrapBatchTools(fun, reg = NULL,
  clusterFunction = batchtools::makeClusterFunctionsInteractive(),
  resources = NULL)
}
\arguments{
\item{fun}{}

\item{reg}{}

\item{clusterFunction}{}

\item{resources}{}
}
\value{
a function
}
\description{
wrapFunction to call a function through batchtools
}
