% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{wrapFunctionParallel}
\alias{wrapFunctionParallel}
\title{Parallelized Function Evaluation Wrapper}
\usage{
wrapFunctionParallel(fun, cl = NULL, nCores = NULL)
}
\arguments{
\item{fun}{the function that shall be evaluated in parallel}

\item{cl}{Optional, an existing computation cluster}

\item{nCores}{Optional, amount of cores to use for creating a new computation cluster. Default is all cores.}
}
\value{
numeric vector, result of the parallelized evaluation
}
\description{
This is a simple wrapper that turns a function of type \code{y=f(x)}, where x is a vector and y is a scalar,
into a function that accepts and returns matrices, as required by \code{\link{spot}}.
While doing so, the wrapper will use the parallel package in order to parallelize the execution of each function
evaluation. This function will create a computation cluster if no cluster is specified and there is no
default cluster setup!
}
