% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{wrapSystem_parseMatrixFromString}
\alias{wrapSystem_parseMatrixFromString}
\title{wrapSystem_parseMatrixFromString}
\usage{
wrapSystem_parseMatrixFromString(str)
}
\arguments{
\item{str}{String that was recieved from the command line}
}
\value{
parsed matrix
}
\description{
Create an R-Matrix from a String that was passed via the command line
}
