% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{wrapSystem_parseMatrixToString}
\alias{wrapSystem_parseMatrixToString}
\title{wrapSystem_parseMatrixToString}
\usage{
wrapSystem_parseMatrixToString(m)
}
\arguments{
\item{m}{a matrix}
}
\value{
parsed string
}
\description{
Create a String that can be passed via the command line from an R-Matrix
}
\keyword{internal}
