% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{wrapSystemCommand}
\alias{wrapSystemCommand}
\title{wrapSystemCommand}
\usage{
wrapSystemCommand(systemCall)
}
\arguments{
\item{systemCall}{String that calls the command line script.}
}
\value{
callable function for SPOT
}
\description{
Optimize parameters for a script that is accessible via Command Line
}
\examples{
exampleScriptLocation <- system.file("consoleCallTrialScript.R",package = "SPOT")
f <- wrapSystemCommand(paste("Rscript", exampleScriptLocation))
spot(,f,c(1,1),c(100,100))
}
