\name{writeRoiFileForRepeats}
\alias{writeRoiFileForRepeats}
\title{write ROI for repeats...}
\usage{writeRoiFileForRepeats(workingDir="~/workspace/SvnFiwaSoma.d/trunk/doc/OwnPublicationsAndTalks.d/Bart10t.d/Experiments.d",
    bstFile="lmSann02.bst", roiInFile="lmSann02.roi", roiOutFile="lmSann03.roi")}
\description{write ROI for repeats}
\arguments{\item{workingDir}{Working directory as string}
\item{bstFile}{best file}
\item{roiInFile}{ROI in file}
\item{roiOutFile}{ROI out file}}
