% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cyclone.R
\name{calculationBarthMuschelknautz}
\alias{calculationBarthMuschelknautz}
\title{Cyclone Simulation: Barth/Muschelknautz}
\usage{
calculationBarthMuschelknautz(cyclone, fluid, xmean, delta)
}
\arguments{
\item{cyclone}{list of cyclone's geometrical parameters}

\item{fluid}{list of fluid parameters}

\item{xmean}{vector of middle points for the intervals of the particle size distribution}

\item{delta}{vector, amount of dust (percentage value) in each interval}
}
\value{
returns a vector, first element is the efficiency, second element is the pressure drop
}
\description{
Calculate cyclone collection efficiency according to Barth/Muschelknautz.
}
\keyword{internal}
