% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.R
\name{evaluateModel}
\alias{evaluateModel}
\title{Evaluate Model}
\usage{
evaluateModel(object, onlyY = F)
}
\arguments{
\item{object}{fit created by a modeling function, e.g., \code{\link{buildRandomForest}}}
}
\value{
a function in the style of \code{y=f(x)}, which uses the fitted object to predict \code{y} for sample \code{x}.
}
\description{
This function produces an objective function with y=f(x) from a provided model fit.
Important note: this function expects \code{predict(object,newdata)} to return
a list. The \code{object$target} parameter is a string that determins which list item
is returned by the created function. If not set (NULL), \code{object$target} is set to \code{"y"}.
}
\keyword{internal}
