% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotPredictForrester.R
\name{forrRegPredictor}
\alias{forrRegPredictor}
\title{Predict Forrester Model}
\usage{
forrRegPredictor(x, fit, pred.all = FALSE)
}
\arguments{
\item{x}{design matrix to be predicted}

\item{fit}{Kriging model (settings and parameters) of class forr}

\item{pred.all}{if TRUE return all (RMSE and prediction, in a dataframe), else return only prediction}
}
\value{
Returned value is dependent on the setting of \code{pred.all}\cr
TRUE: data.frame with columns f (function values) and s (RMSE)\cr
FALSE: vector of function values only
}
\description{
This function is for backwards compatibiltiy only. See new predictor: \code{\link{predict.forr}}
}
\seealso{
\code{\link{predict.forr}}
}

