% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funContinuous.R
\name{funRosen}
\alias{funRosen}
\title{funRosen}
\usage{
funRosen(x)
}
\arguments{
\item{x}{matrix of points to evaluate with the function. Rows for points and columns for dimension.}
}
\value{
1-column matrix with resulting function values
}
\description{
Rosenbrock Test Function
}
\examples{
x1 <- matrix(c(1,1),1,)
funRosen(x1)

}
\references{
More', J. J., Garbow, B. S., & Hillstrom, K. E. (1981).
Testing unconstrained optimization software.
\emph{ACM Transactions on Mathematical Software (TOMS)}, \emph{7}(1), 17-41.
\doi{10.1145/355934.355936}

Rosenbrock, H. (1960).
An automatic method for finding the greatest or least value of a function.
\emph{The Computer Journal}, \emph{3}(3), 175-184.
\doi{10.1093/comjnl/3.3.175}
}
