% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sphere.R
\name{funSphere}
\alias{funSphere}
\title{Sphere Test Function}
\usage{
funSphere(x)
}
\arguments{
\item{x}{matrix of points to evaluate with the sphere function. Rows for points and columns for dimension.}
}
\value{
1-column matrix with resulting function values
}
\description{
Sphere Test Function
}
\examples{
funSphere(matrix(runif(18),,3))
}
