% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/infillGetFullPrediction.R
\name{infillGetFullPrediction}
\alias{infillGetFullPrediction}
\title{infillGetFullPrediction}
\usage{
infillGetFullPrediction(predictionList, model)
}
\arguments{
\item{predictionList}{}

\item{model}{}
}
\value{
list, full model predictio
}
\description{
Wrapper to get the full model Prediction instead of just one target for plotting purposes
}
\keyword{internal}
