% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildKrigingForrester.R
\name{normalizeMatrix}
\alias{normalizeMatrix}
\title{Normalize design}
\usage{
normalizeMatrix(x, ymin, ymax)
}
\arguments{
\item{x}{design matrix in input space}

\item{ymin}{minimum vector of normalized space}

\item{ymax}{maximum vector of normalized space}
}
\value{
normalized design matrix
}
\description{
Normalize design by using minimum and maximum of the design values for input space. Supportive function for Kriging model, not to be used directly.
}
\seealso{
\code{\link{buildKriging}}
}
\keyword{internal}
