% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildTree.R
\name{plot.spotTreeModel}
\alias{plot.spotTreeModel}
\title{Plot rpart tree model}
\usage{
\method{plot}{spotTreeModel}(x, ...)
}
\arguments{
\item{x}{tree model (settings and parameters) of class \code{spotTreeModel}.}

\item{...}{parameters passed to plot function}
}
\description{
Simple plot of the model produced by \code{\link{buildTreeModel}}.
}
\details{
A nicer plot can be generated with the \code{plotnice()} function
from the \code{SPOTMisc} package.
}
\keyword{internal}
