% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/buildKrigingDACE.R, R/buildKrigingForrester.R
\name{print.dace}
\alias{print.dace}
\alias{print.kriging}
\title{Print Function DACE Kriging}
\usage{
\method{print}{dace}(x, ...)

\method{print}{kriging}(x, ...)
}
\arguments{
\item{x}{fit returned by \code{\link{buildKriging}}.}

\item{...}{additional parameters}
}
\description{
Print information about a DACE Kriging fit, as produced by \code{\link{buildKrigingDACE}}.

Print information about a Forrester Kriging fit, as produced by \code{\link{buildKriging}}.
}
\keyword{internal}
