\name{print.spotConfig}
\alias{print.dace}
\alias{print.forr}
\alias{print.spotConfig}
\title{Print function for spotConfig class}
\usage{
  \method{print}{spotConfig} (x, ...)

  \method{print}{dace} (x, ...)

  \method{print}{forr} (x, ...)
}
\arguments{
  \item{x}{spotConfig}

  \item{...}{additional parameters}

  \item{x}{fit returned by \code{\link{forrBuilder}}.}

  \item{...}{additional parameters}

  \item{x}{fit returned by \code{\link{forrBuilder}}.}

  \item{...}{additional parameters}
}
\description{
  Print function to summarize a spotConfig.

  Print information about a DACE Kriging fit, as produced
  by \code{\link{daceBuilder}}.

  Print information about a Forrester Kriging fit, as
  produced by \code{\link{forrBuilder}}.
}
\keyword{internal}

