% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/selectAll.R
\name{selectAll}
\alias{selectAll}
\title{selectAll}
\usage{
selectAll(x, y, control)
}
\arguments{
\item{x}{matrix of design points}

\item{y}{matrix of function values}

\item{control}{list of controls}
}
\value{
This function returns a list with:
\describe{
	\item{\code{x}}{matrix of selected design points.}
	\item{\code{y}}{ matrix of selected function values.}
}
}
\description{
Select all Design Points. 
Returns the whole set of proposed design points.
It is a dummy function, used as the default for more sophisticated selection procedures.
}
\keyword{internal}
