% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot.R
\docType{package}
\name{SPOT-package}
\alias{SPOT}
\alias{SPOT-package}
\title{Sequential Parameter Optimization Toolbox in R}
\description{
Sequential Parameter Optimization Toolbox in R
}
\details{
SPOT is a package for R, using statistic models to find
optimal parameters for optimization algorithms. SPOT is a very flexible and 
user oriented tool box for parameter optimization. The flexibility has its 
price: to fully use all the possibilities of flexibility the user is requested 
to look at a large number of spot-parameters to change. The good news is, that 
some defaults are given that already work perfectly well for 90 percent of the users.

\tabular{ll}{
Package: \tab SPOT\cr
Type: \tab Package\cr
Version: \tab 1.1.0\cr
Date: \tab 2016-06-08\cr
License: \tab GPL (>= 2)\cr
LazyLoad: \tab yes\cr
}
}
\section{Acknowledgments}{

This work has been partially supported by the Federal Ministry of Education
and Research (BMBF) under the grants CIMO (FKZ 17002X11) and
MCIOP (FKZ 17N0311).
}

\section{Maintainer}{

Martin Zaefferer \email{martin.zaefferer@gmx.de}
}
\author{
Thomas Bartz-Beielstein \email{thomas.bartz-beielstein@fh-koeln.de} with contributions from: J. Ziegenhirt, W.
   Konen, O. Flasch, M. Friese, P. Koch, M. Zaefferer, B. Naujoks, M. Friese
}
\seealso{
Main interface functions are \code{\link{spot}} and \code{\link{spotOptim}}.
Also, a graphical interface can be used with \code{\link{spotGui}}
}
\keyword{package}

