\name{SPOT-package}
\title{Sequential Parameter Optimization Toolbox in R}
\description{Sequential Parameter Optimization Toolbox in R}
\details{\tabular{ll}{
Package: \tab SPOT\cr
Type: \tab Package\cr
Version: \tab 0.1.1375\cr
Date: \tab 03.05.2011\cr
License: \tab GPL (>= 3)\cr
LazyLoad: \tab yes\cr
}

SPOT is a package for R, using statistic models to find
optimal parameters for optimization algorithms. SPOT is a very flexible and 
user oriented tool box for parameter optimization. The flexibility has its 
price: to fully use all the possibilities of flexibility the user is requested 
to look at a large number of spot-parameters to change. The good news is, that 
some defaults are given that already work perfectly well for 90 percent of the users.}
\alias{SPOT}
\docType{package}
\author{Thomas Bartz-Beielstein \email{thomas.bartz-beielstein@fh-koeln.de} with contributions from: J. Ziegenhirt, W.
Konen, O. Flasch, P. Koch, M. Zaefferer}
\references{\url{http://www.gm.fh-koeln.de/campus/personen/lehrende/thomas.bartz-beielstein/00489/} \cr
\url{http://www.springer.com/3-540-32026-1}}
\keyword{package}
\seealso{Main interface functions are \code{\link{spot}} and \code{\link{spotOptim}}, 
also a graphical interface can be used with \code{\link{spotGui}}}
