% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPOT-package.R
\docType{package}
\name{SPOT-package}
\alias{SPOT}
\alias{SPOT-package}
\title{SPOT: SPatial Omnibus Test (SPOT)}
\description{
Perform a SPatial Omnibus Test (SPOT) to relate single cell spatial data to patient outcomes.
}
\author{
\strong{Maintainer}: Sarah Samorodnitsky \email{ssamorod@fredhutch.org} (\href{https://orcid.org/0000-0001-7909-0245}{ORCID})

}
\keyword{internal}
