% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SPOT.R
\name{spot}
\alias{spot}
\title{SPatial Omnibus Test (SPOT)}
\usage{
spot(
  data,
  radii,
  outcome,
  censor = NULL,
  model.type,
  use.K = TRUE,
  homogeneous = TRUE,
  adjustments = NULL,
  marked = FALSE,
  cell.type = NULL,
  K.diff = FALSE,
  pick.roi = "all",
  agg.func = NULL,
  seed = 12345,
  nlarge = 10000,
  prop.nonzero = 0.2,
  print.progress = TRUE
)
}
\arguments{
\item{data}{The dataset of single cell spatial data. Must include column names
PID, id, x, y, and type if \code{marked}. PID corresponds to the patient IDs
which should be replicated across ROIs for each patient. id corresponds to the
image ID within a patient. x and y give the coordinates of each cell. type (if given)
is a cell type label. Should also contain an outcome column which will be specified in
the \code{outcome} parameter.}

\item{radii}{Vector of length \eqn{P} giving the radii at which to evaluate the
spatial summary measure. The first value must be zero.}

\item{outcome}{String specifying the column name in \code{data} containing the outcome.
Must be a survival outcome, a binary outcome, or a continuous outcome.}

\item{censor}{String specifying the column name in \code{data} containing the censoring indicator.
Leave this parameter \code{NULL} if the outcome is not survival.}

\item{model.type}{String specifying which predictive model to use. The options are: "survival" for a
Cox proportional hazards model, or "logistic" for a logistic regression model, or "linear" for a
standard linear model.}

\item{use.K}{Boolean. If \code{TRUE}, calculates Ripley's K for each radius and
sample. If \code{FALSE}, calculates Besag's L for each radius and sample.}

\item{homogeneous}{Boolean. If \code{TRUE}, calculates the homogeneous versions of
Ripley's K or Besag's L. If \code{FALSE}, calculates the inhomogeneous versions of
these summaries.}

\item{adjustments}{String(s) specifying the column names in \code{data} of covariates
to adjust for in the predictive model. May be left \code{NULL}.}

\item{marked}{Boolean. If \code{TRUE}, subsets the data to consider only cells with the
labels given in the \code{cell.type} parameter. If \code{FALSE}, considers all detected cells
included in \code{data}.}

\item{cell.type}{The cell type labels to consider. Can be one string, e.g. \code{cell.type = "CD4 T cell"}, or
a vector of two cell types, e.g. \code{cell.type = c("CD4 T cell", "CD8 T cell")}. If \code{!marked}, leave
this parameter as \code{NULL}.}

\item{K.diff}{Boolean. If \code{TRUE}, returns the difference between the empirical and theoretical values of
Ripley's K or Besag's L. If \code{FALSE}, returns the empirical value of the spatial summary.}

\item{pick.roi}{String specifying how to select among ROIs if multiple are available per sample. Default is \code{all}
which should be used if only one image is available per person. If multiple are available, specify this parameter
as \code{"average"} to average the spatial summaries across ROIs within each radius, \code{random} to randomly select
an ROI for each sample, or \code{custom} to specify a custom user-provided function to aggregate spatial summaries.
If \code{pick.roi = custom}, user must supply function using argument \code{agg.func}.}

\item{agg.func}{If \code{pick.roi = custom}, user must supply a function to aggregate spatial summaries across
multiple ROIs.}

\item{seed}{Specify a seed to use if \code{pick.roi = "random"}.}

\item{nlarge}{Parameter for the \code{Kest} and \code{Lest} function for efficiency. If there are more cells
than \code{nlarge}, will apply a border correction for edge effects instead of the isotropic edge correction.
Default is 10000.}

\item{prop.nonzero}{How many spatial summaries should be non-zero in order for the corresponding radius to be
included in the predictive model? Specify as a proportion of the total number of samples. Default is 0.2.}

\item{print.progress}{Boolean. Should progress in calculating the spatial summary across samples be printed?
If \code{TRUE}, prints a progress bar.}
}
\value{
Returns a list with the following elements:
\item{overall.pval}{The overall p-value resulting from the Cauchy combination test characterizing the
strength of association between the spatial summary and the outcome across radii.}
\item{Kr.df}{The matrix of spatial summaries across samples and radii. If multiple ROIs were person and \code{pick.roi = "average" or "random"},
returns only the averaged spatial summaries or the summaries for the randomly-selected ROIs.}
\item{pval.df}{P-values for the association between the spatial summary at each radius and the outcome. }
\item{pids.images}{The PIDs and the corresponding image IDs within each patient. }
\item{images.missing.cells}{How many images were missing cells altogether?}
\item{n.images.association}{How many images were used in the predictive model?}
\item{arguments}{Returns the arguments specified in the function call.}
}
\description{
Perform a SPatial Omnibus Test (SPOT) to relate a spatial summary of each
image to patient outcomes.
}
