\name{spotAlgEsBananaGradient}
\alias{spotAlgEsBananaGradient}
\title{Gradient Function of 'spotAlgEsBanana'...}
\usage{spotAlgEsBananaGradient(x)}
\description{Gradient Function of 'spotAlgEsBanana'}
\details{This is a fitness function that could be used for the ES which is an example algorithm to be optimized by SPOT.}
\value{number \code{res} \cr
- \code{res} is the function value of the corresponding vector \code{x}}
\references{\code{\link{spotAlgEsBanana}} \code{\link{spotAlgEsF}}}
\arguments{\item{x}{vector to be evaluated by the fitness function}}
