% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimES.R
\name{spotAlgEsDominantReco}
\alias{spotAlgEsDominantReco}
\title{spotAlgEsDominantReco}
\usage{
spotAlgEsDominantReco(parents, rhoVal, dimension, nSigma, objType = "obj")
}
\arguments{
\item{parents}{Parent individuals}

\item{rhoVal}{number of parents involved in the procreation of an offspring}

\item{dimension}{number of dimensions}

\item{nSigma}{number of standard deviations}

\item{objType}{string, default is "obj"}
}
\description{
Recombination function for the Evolution Strategy.
}
\seealso{
\code{\link{spotAlgEs}} \code{\link{spotAlgEsInterRecoBeSw02}}  \code{\link{spotAlgEsInterReco}}
}
\keyword{internal}
