\name{spotAlgEsDominantReco}
\alias{spotAlgEsDominantReco}
\title{spotAlgEsDominantReco...}
\usage{spotAlgEsDominantReco(parents, rhoVal, dimension, nSigma,
    objType="obj")}
\description{spotAlgEsDominantReco}
\details{Recombination function for the ES}
\value{\code{rObject}}
\arguments{\item{parents}{Parent individuals}
\item{rhoVal}{number of parents involved in the procreation of an offspring}
\item{dimension}{number of dimensions}
\item{nSigma}{number of standard deviations}
\item{objType}{string, default is "obj"}}
