% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimES.R
\name{spotAlgEsGetSuccessRate}
\alias{spotAlgEsGetSuccessRate}
\title{get Success Rate}
\usage{
spotAlgEsGetSuccessRate(gen, pop)
}
\arguments{
\item{gen}{Generation number}

\item{pop}{Population to be evaluated (only the generation \code{gen} will be evaluated)}
}
\value{
number \cr
- the success rate of the given population
}
\description{
This function determines the success rate.
}
\seealso{
\code{\link{spotAlgEs}}
}
\keyword{internal}
