\name{spotAlgEsGetSuccessRate}
\alias{spotAlgEsGetSuccessRate}
\title{get Success Rate...}
\usage{spotAlgEsGetSuccessRate(gen, pop)}
\description{get Success Rate}
\details{This function is a threshold function, evaluating the success rate.}
\value{number \cr
- the success rate of the given population}
\references{\code{\link{spotAlgEs}} \code{\link{spotAlgStartEs}}}
\arguments{\item{gen}{Generation number}
\item{pop}{Population to be evaluated (only the generation \code{gen} will be evaluated)}}
