\name{spotAlgEsGetSuccessRate}
\alias{spotAlgEsGetSuccessRate}
\title{get Success Rate...}
\usage{spotAlgEsGetSuccessRate(gen, pop)}
\description{get Success Rate}
\details{This function is a threshold function, evaluating the success rate.}
\value{number \code{succ} \cr
- \code{succ} is the success rate}
\arguments{\item{gen}{Generation number}
\item{pop}{Population to be evaluated}}
