% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimES.R
\name{spotAlgEsHps}
\alias{spotAlgEsHps}
\title{Termination hps}
\usage{
spotAlgEsHps(xk, xOpt)
}
\arguments{
\item{xk}{current best value of the optimization run}

\item{xOpt}{target value of the optimization run}
}
\value{
\code{boolean} \cr
- TRUE as long as the current value has not yet reached its limit. Once the
given termination criterion is reached the function returns FALSE.
}
\description{
Termination function for the ES. Terminates the ES at a given target value.
}
\keyword{internal}
