\name{spotAlgEsHps}
\alias{spotAlgEsHps}
\title{Termination hps...}
\usage{spotAlgEsHps(xk, xOpt)}
\description{Termination hps}
\details{Termination function for the ES}
\value{\code{boolean}}
\arguments{\item{xk}{current best value of the optimization run}
\item{xOpt}{target value of the optimization run}}
