% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimES.R
\name{spotAlgEsIndividualInitial}
\alias{spotAlgEsIndividualInitial}
\title{Individual Initialization}
\usage{
spotAlgEsIndividualInitial(
  s,
  dimension,
  n,
  noise = 0,
  fName,
  gen,
  low = -1,
  high = 1,
  des,
  ...
)
}
\arguments{
\item{s}{sigma, step size}

\item{dimension}{number of target function dimension}

\item{n}{n, number of diff. step sizes}

\item{noise}{noise to be added to fitness value}

\item{fName}{target function}

\item{gen}{generation}

\item{low}{lower limit}

\item{high}{upper limit}

\item{des}{des scaling for placement between low and high}

\item{...}{additional parameters to be passed on to \code{fName}}
}
\value{
numeric vector \cr
- contains x value, sigma value, real fitness value, fitness with noise, and generation number
}
\description{
Creates a new Individual for the Evolution Strategy implemented in SPOT.
}
\seealso{
\code{\link{spotAlgEs}}
}
\keyword{internal}
