\name{spotAlgEsIndividualInitial}
\alias{spotAlgEsIndividualInitial}
\title{Individual Initial...}
\usage{spotAlgEsIndividualInitial(s, dimension, n, noise=0, fName, gen,
    low=-1, high=1, des)}
\description{Individual Initial}
\value{number \code{res} \cr
- \code{res} is the value of the corresponding \code{x} vector}
\arguments{\item{s}{sigma}
\item{n}{n}
\item{dimension}{number of target function dimension}
\item{noise}{noise to be added to fitness value}
\item{fName}{function name}
\item{gen}{generation}
\item{low}{lower limit}
\item{high}{upper limit}
\item{des}{des}}
