% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimES.R
\name{spotAlgEsInitParentPop}
\alias{spotAlgEsInitParentPop}
\title{Initialize Parent Population}
\usage{
spotAlgEsInitParentPop(sigmaInit, dimension, nSigma, noise, fName, gen,
  low, high, mue, ...)
}
\arguments{
\item{sigmaInit}{initial sigma value (standard deviation)}

\item{dimension}{number of target function dimension}

\item{nSigma}{number of standard deviations}

\item{noise}{noise to be added to fitness value}

\item{fName}{target function}

\item{gen}{generation}

\item{low}{lower limit}

\item{high}{upper limit}

\item{mue}{number of parents in the ES}

\item{...}{additional parameters to be passed on to \code{fName}}
}
\value{
matrix \cr
- holds the parent population created by this function
}
\description{
Creates initial parent population
}
\seealso{
\code{\link{spotAlgEs}} \code{\link{spotAlgEsIndividualInitial}}
}
\keyword{internal}
