% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimES.R
\name{spotAlgEsMarriageWithReplace}
\alias{spotAlgEsMarriageWithReplace}
\title{Marriage with replace}
\usage{
spotAlgEsMarriageWithReplace(pop, rhoVal)
}
\arguments{
\item{pop}{Population}

\item{rhoVal}{number of parents involved in the procreation of an offspring}
}
\value{
\code{pop} \cr
- \code{pop} Population
}
\description{
Recombination function for the Evolution Strategy.
}
\seealso{
\code{\link{spotAlgEs}} \code{\link{spotAlgEsMarriageWithReplace}}
}
\keyword{internal}
