\name{spotAlgEsQuickTest}
\alias{spotAlgEsQuickTest}
\title{ES Quick Test...}
\usage{spotAlgEsQuickTest(mue=2, nu=2, tau=1)}
\description{ES Quick Test}
\details{This is a quick test function for the ES.}
\references{\code{\link{spotAlgEs}} \code{\link{spotAlgStartEs}}}
\arguments{\item{mue}{value of mue (number of parents) to be used with the ES default is 2}
\item{nu}{value of nu to be used with the ES default is 2}
\item{tau}{value of tau (learning parameter) to be used with the ES default is 1}}
