% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimES.R
\name{spotAlgEsSelection}
\alias{spotAlgEsSelection}
\title{spotAlgEsSelection}
\usage{
spotAlgEsSelection(parentPop, offspringPop, sel, gen, iter, maxIter)
}
\arguments{
\item{parentPop}{Parent population}

\item{offspringPop}{Offspring population}

\item{sel}{number of individuals to be selected}

\item{gen}{generation number}

\item{iter}{current iteration number}

\item{maxIter}{maximum iteration number}
}
\description{
Selection function for the ES
}
\keyword{internal}
