\name{spotAlgEsSphere}
\alias{spotAlgEsSphere}
\title{Sphere Function...}
\usage{spotAlgEsSphere(x)}
\description{Sphere Function}
\details{This is a fitness function that could be used for the ES.}
\value{number \code{res} \cr
- \code{res} is the value of the corresponding \code{x} vector}
\references{\code{\link{spotAlgEsSphere1}}}
\arguments{\item{x}{vector to be evaluated by the fitness function}}
