% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimES.R
\name{spotAlgEsStratMutation}
\alias{spotAlgEsStratMutation}
\title{spotAlgEsStratMutation}
\usage{
spotAlgEsStratMutation(strat, tau0, tau, sigmaRestart, sigmaInit)
}
\arguments{
\item{strat}{Strategy parameter (sigma), can be a vector or a single number}

\item{tau0}{the global/general step size multiplier for self adaption}

\item{tau}{the step size multiplier for self adaption in each dimension}

\item{sigmaRestart}{sigma values are reset to initial values when a uniformly distributed random number is smaller than sigmaRestart}

\item{sigmaInit}{initial sigma value}
}
\value{
number \code{s} \cr
- \code{s} is the new sigma value
}
\description{
Mutation function for the ES strategy parameter sigma.
}
\keyword{internal}
