% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimES.R
\name{spotAlgEsTermination}
\alias{spotAlgEsTermination}
\title{Termination}
\usage{
spotAlgEsTermination(it, maxIt, ge, maxGe, xk, xOpt = NULL)
}
\arguments{
\item{it}{iteration (function evaluations)}

\item{maxIt}{Maximum number of iterations (function evaluations)}

\item{ge}{generation}

\item{maxGe}{Maximum number of generations}

\item{xk}{current best value of the optimization run}

\item{xOpt}{target value of the optimization run}
}
\value{
\code{boolean} \cr
- TRUE as long as the current value has not yet reached its limit. Once the
given termination criterion is reached the function returns FALSE.
}
\description{
Handles the termination functions for the ES.
}
\seealso{
\code{\link{spotAlgEs}} \code{\link{spotAlgEsHps}}
}
\keyword{internal}
