% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spotCreateDesignFrF2.R
\name{spotCreateDesignFrF2}
\alias{spotCreateDesignFrF2}
\title{spotCreateDesignFrF2}
\usage{
spotCreateDesignFrF2(spotConfig, noDesPoints = NaN, repeats = NaN)
}
\arguments{
\item{spotConfig}{list of spot settings}

\item{noDesPoints}{is obsolete for this type of design, it has no influence. The number of points is fixed.}

\item{repeats}{is obsolete for this type of design, it has no influence.}
}
\value{
matrix \code{M} \cr
- \code{M} has \code{dimension} columns and \code{noDesPoints} rows
}
\description{
Uses the function FrF2 from the FrF2-package to generate a (fractional) factorial design.
The factorial design is augmented with the augment.ccd function from DoE.wrapper package.
The dimension is determined from the number of rows of the .roi - file
(each row in the roi file defines a variable).
}
\seealso{
\code{\link{spotCreateDesignBasicDoe}}, \code{\link{spotCreateDesignLhd}},
\code{\link{spotCreateDesignLhs}}, \code{\link{spotCreateDesignLhsOpt}}
}

