% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spotEnsembleFunctions.R
\name{spotEnsembleMinimum}
\alias{spotEnsembleMinimum}
\title{spotEnsembleMinimum}
\usage{
spotEnsembleMinimum(x)
}
\arguments{
\item{x}{result structure: a list with element \code{Y} for predicted values and \code{fits} for model fit list}
}
\value{
the summed
}
\description{
Minn over all points.
}
\keyword{internal}

