% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot.R
\name{spotFillControlList}
\alias{spotFillControlList}
\title{Fill in some values for the control list. Internal use only.}
\usage{
spotFillControlList(controlList, dimension)
}
\arguments{
\item{controlList}{list of controls, see \code{\link{spotControl}}.}

\item{dimension}{dimension of the optimization problem. See \code{\link{spotControl}}.}
}
\value{
a list
}
\description{
Fill in some values for the control list. Internal use only.
}
\keyword{internal}
