% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotPredictHelpFunctions.R
\name{spotForceFactorDesign}
\alias{spotForceFactorDesign}
\title{Force into Factors: Design
 
Forces the type of parameters in the design to be predicted to factors, if type[i] is specified to "FACTOR"}
\usage{
spotForceFactorDesign(design, type, xNames, lvls)
}
\arguments{
\item{design}{prediction locations}

\item{type}{string of type information, factors should be  "FACTOR", rest is ignored}

\item{xNames}{names of the parameters in x}

\item{lvls}{levels to be used when forcing the parameter values to factor type}
}
\value{
returns the design variable, but now with factors where necessary.
}
\description{
Force into Factors: Design
 
Forces the type of parameters in the design to be predicted to factors, if type[i] is specified to "FACTOR"
}
\keyword{internal}

