% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotPredictHelpFunctions.R
\name{spotForceFactorX}
\alias{spotForceFactorX}
\title{Force into Factors: X
 
Forces the type of parameters in the sample locations x to factors, if type[i] is specified to "FACTOR"}
\usage{
spotForceFactorX(x, type, xNames)
}
\arguments{
\item{x}{sample locations}

\item{type}{string of type information, factors should be  "FACTOR", rest is ignored}

\item{xNames}{names of the parameters in x}
}
\value{
returns a list with component \code{x} (data frame x, now with factors) \code{levels} (list of levels of the corresponding factors).
}
\description{
Force into Factors: X
 
Forces the type of parameters in the sample locations x to factors, if type[i] is specified to "FACTOR"
}
\keyword{internal}

