\name{spotGetAllRepeats}
\alias{spotGetAllRepeats}
\title{Get ALL Repeats Function...}
\usage{spotGetAllRepeats(configList)}
\description{Get ALL Repeats Function}
\details{Gets repeats of all configs in the list}
\value{list \code{all} \cr
- \code{all} contains all repeats of the given config list}
\arguments{\item{configList}{config list}}
