% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spotStatistics.R
\name{spotGetAllRepeats}
\alias{spotGetAllRepeats}
\title{Get ALL Repeats Function}
\usage{
spotGetAllRepeats(configList)
}
\arguments{
\item{configList}{config list}
}
\value{
list \code{all} \cr
- \code{all} contains all repeats of the given config list
}
\description{
Gets repeats of all configs in the list
}
\keyword{internal}

