% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spotStatistics.R
\name{spotGetRepeats}
\alias{spotGetRepeats}
\title{Get Repeats Function}
\usage{
spotGetRepeats(configNumber, configList)
}
\arguments{
\item{configNumber}{one configuration number}

\item{configList}{list of configuration numbers}
}
\value{
number \code{repList} \cr
-  \code{repList} contains the number of repeats of the given config
}
\description{
Helper Function for Spot's stats, evaluating the number of repeats
for a given parameter configuration
}
\keyword{internal}

