\name{spotGetRepeats}
\alias{spotGetRepeats}
\title{Get Repeats Function...}
\usage{spotGetRepeats(configNumber, configList)}
\description{Get Repeats Function}
\details{Helper Function for Spot's stats, evaluating the number of repeats
for a given parameter configuration}
\value{number \code{repList} \cr
-  \code{repList} contains the number of repeats of the given config}
\arguments{\item{configNumber}{one configuration number}
\item{configList}{list of configuration numbers}}
