\name{spotGui}
\alias{spotGui}
\title{Start the SPOT GUI...}
\description{Start the SPOT GUI}
\details{This function starts the graphical user interface for SPOT, which
is based on java. The .jar File started with this function is
in the directory where SPOT is installed. 

Java needs to be installed to use the gui.}
\references{\code{\link{SPOT}}}
