\name{spotHlpF.norm}
\alias{spotHlpF.norm}
\title{Helper function for transformations/normalizing...}
\usage{spotHlpF.norm(x)}
\description{Helper function for transformations/normalizing}
\details{Transforms natural to coded variables, used by spotPredictLm}
\value{vector \code{y} \cr
- \code{y} is the transformed/normalized version of \code{x}}
\arguments{\item{x}{parameter vector}}
