% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spotStatistics.R
\name{spotHlpF.norm}
\alias{spotHlpF.norm}
\title{Helper function for transformations/normalizing}
\usage{
spotHlpF.norm(x)
}
\arguments{
\item{x}{parameter vector}
}
\value{
vector \code{y} \cr
- \code{y} is the transformed/normalized version of \code{x}
}
\description{
Transforms natural to coded variables, used by spotPredictLm
}
\keyword{internal}

