\name{spotHlpF.norminv}
\alias{spotHlpF.norminv}
\title{Helper function for transformations/normalizing...}
\usage{spotHlpF.norminv(x, z)}
\description{Helper function for transformations/normalizing}
\details{Transforms natural to coded variables}
\value{vector \code{y} \cr
- \code{y} is the transformed/normalized version of \code{x}}
\arguments{\item{x}{parameter vector}
\item{z}{vector (zmin, zmax)}}
