\name{spotHlpF.norminv}
\alias{spotHlpF.norminv}
\title{Helper function for transformations/normalizing}
\usage{
  spotHlpF.norminv(x, z)
}
\arguments{
  \item{x}{parameter vector}

  \item{z}{vector (zmin, zmax)}
}
\value{
  vector \code{y} \cr - \code{y} is the
  transformed/normalized version of \code{x}
}
\description{
  Transforms natural to coded variables
}
\keyword{internal}

