% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotPredictForrester.R
\name{spotNormalizeMatrix}
\alias{spotNormalizeMatrix}
\title{Normalize design}
\usage{
spotNormalizeMatrix(x, ymin, ymax, xmin = NULL, xmax = NULL)
}
\arguments{
\item{x}{design matrix in input space}

\item{ymin}{minimum vector of normalized space}

\item{ymax}{maximum vector of normalized space}
}
\value{
normalized design matrix
}
\description{
Normalize design by using minimum and maximum of the design values for input space
}
\seealso{
\code{\link{spotPredictForrester}} \code{\link{forrBuilder}}
 \code{\link{spotNormalizeMatrix2}}
\code{\link{spotReverseNormalizeMatrix}}
}
\keyword{internal}

