\name{spotParetoOptMulti}
\alias{spotParetoOptMulti}
\title{DEPRECATED}
\usage{
  spotParetoOptMulti(startPoint, spotConfig)
}
\arguments{
  \item{startPoint}{see \code{\link{spotModelParetoOptim}}}

  \item{spotConfig}{see \code{\link{spotModelParetoOptim}}}
}
\description{
  see \code{\link{spotModelParetoOptim}}
}
\keyword{internal}

