% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotModelParetoOptim.R
\name{spotParetoOptMulti}
\alias{spotParetoOptMulti}
\title{DEPRECATED
 
see \code{\link{spotModelParetoOptim}}}
\usage{
spotParetoOptMulti(startPoint, spotConfig)
}
\arguments{
\item{startPoint}{see \code{\link{spotModelParetoOptim}}}

\item{spotConfig}{see \code{\link{spotModelParetoOptim}}}
}
\description{
DEPRECATED
 
see \code{\link{spotModelParetoOptim}}
}
\keyword{internal}

