% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSeverity.R
\name{spotPower}
\alias{spotPower}
\title{spotPower}
\usage{
spotPower(alpha, mu0, mu1, n, sigma)
}
\arguments{
\item{alpha}{description of alpha}

\item{mu0}{description of mu0}

\item{mu1}{description of mu1}

\item{n}{vector length}

\item{sigma}{standart deviation}
}
\value{
description of return value
}
\description{
Calculate power
}
