\name{spotPower}
\alias{spotPower}
\title{spotPower}
\usage{
  spotPower(alpha, mu0, mu1, n, sigma)
}
\arguments{
  \item{alpha}{description of alpha}

  \item{mu0}{description of mu0}

  \item{mu1}{description of mu1}

  \item{n}{vector length}

  \item{sigma}{standard deviation}
}
\value{
  description of return value
}
\description{
  short description
}
\details{
  details
}
\seealso{
  \code{\link{spotPlotPower}} \code{\link{spotCompare}}
}
\keyword{internal}

