% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotModelOptim.R
\name{spotPredictOptMulti}
\alias{spotPredictOptMulti}
\title{DEPRECATED
 
see \code{\link{spotModelOptim}}}
\usage{
spotPredictOptMulti(startPoint, spotConfig)
}
\arguments{
\item{startPoint}{see \code{\link{spotModelOptim}}}

\item{spotConfig}{see \code{\link{spotModelOptim}}}
}
\description{
DEPRECATED
 
see \code{\link{spotModelOptim}}
}
\keyword{internal}

