% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spotHelpFunctions.R
\name{spotPrint}
\alias{spotPrint}
\title{Spot Print}
\usage{
spotPrint(set.io.verbosity, io.verbosity, myArg)
}
\arguments{
\item{set.io.verbosity}{\code{spotConfig$io.verbosity} should be passed here. Global flag to drive the \code{io.verbosity} of the program}

\item{io.verbosity}{\code{io.verbosity} for this specified output string}

\item{myArg}{the argument to be printed}
}
\description{
This help function prints the given argument "myArg"
only if user gives the io.verbosity to do so
}
\seealso{
\code{\link{SPOT}} \code{\link{writeLines}}
}
\keyword{internal}

