\name{spotReadRoi}
\alias{spotReadRoi}
\title{Spot Read ROI...}
\usage{spotReadRoi(roiFile, sep)}
\description{Spot Read ROI}
\details{help function spotReadRoi reads region of interest from the .roi-file}
\value{data.frame \code{aroi} \cr
- \code{roi} contains the data from the roi file}
\arguments{\item{roiFile}{this file contains the roi}
\item{sep}{this is used as a column separator}}
