% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotReportMetaDefault.R
\name{spotReportMetaDefault}
\alias{spotReportMetaDefault}
\title{Default Report for Meta Runs}
\usage{
spotReportMetaDefault(spotConfig)
}
\arguments{
\item{spotConfig}{the configuration list of all spot parameters}
}
\description{
Function to generate a simple report for meta runs.
}
\details{
This function draws a scatterplot matrix (based on car), printing it to screen or pdf. 
If the \code{report.io.pdf} setting is TRUE
the graphic is printed to a pdf file (usually named like your .conf file, and placed in the same folder)
if \code{report.io.screen} is set TRUE the graphic is printed to the screen. Both can be FALSE or TRUE
at the same time. If the user does not specify those values, the defaults will be used as shown in
\code{\link{spotGetOptions}}, which means there will be only screen output, and no pdf.
}
\seealso{
\code{\link{SPOT}} \code{\link{spot}} \code{\link{spotStepReport}}
}

