% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spotReportSens.R
\name{spotReportSens}
\alias{spotReportSens}
\title{Sensitivity Report}
\usage{
spotReportSens(spotConfig)
}
\arguments{
\item{spotConfig}{the configuration list of all spot parameters}
}
\description{
Function to generate a report with sensitivity plot.
}
\details{
The sensitivity curves are based on a metamodel which is a random forest with 100 trees 
fitted to the result points from RES-file. The plot contains: 
   x-axis:   ROI for each parameter normalized to [-1,1]
   y-axis:
}
\seealso{
\code{\link{SPOT}} \code{\link{spot}} \code{\link{spotStepReport}}
}

