\name{spotRsm}
\alias{spotRsm}
\title{rsm Model...}
\usage{spotRsm(formula, data)}
\description{rsm Model}
\details{Help function creating a response-surface Regression
this function will be replaced when rsm-package has fixed a bug}
\value{model usd by predict \cr
fix of rsm - Model}
\arguments{\item{formula}{must be a formula}
\item{data}{parameter holding the data for the formula}}
