% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reportSeverity.R
\name{spotSeverity}
\alias{spotSeverity}
\title{spotSeverity}
\usage{
spotSeverity(x0, mu1, n, sigma, alpha)
}
\arguments{
\item{x0}{sample mean value}

\item{mu1}{description}

\item{n}{description}

\item{sigma}{description}

\item{alpha}{description}
}
\value{
description of return value
}
\description{
spotSeverity
}
