% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spot.R
\name{spotStepAutoOpt}
\alias{spotStepAutoOpt}
\title{SPOT Step Auto Opt}
\usage{
spotStepAutoOpt(spotConfig, ...)
}
\arguments{
\item{spotConfig}{the list of all parameters is given, it is forwarded to the call of the report-function
the used parameters of spotConfig are just spotConfig$auto.loop.steps
specifying the number of meta models that should be calculated}

\item{...}{additional parameters to be passed on to target function which is called inside alg.func}
}
\description{
spotStepAutoOpt is the default task called, when spot is started.
}
\details{
The \code{auto} task calls the tasks \code{init} and \code{run} once
and loops \code{auto.loop.steps} times over the steps \code{seq}  and \code{run}
finalising the function with a call of the report function. Instead of \code{auto.loop.steps}
also \code{auto.loop.nevals} can be used as a stopping criterion.
}
\seealso{
\code{\link{SPOT}} \code{\link{spot}} \code{\link{spotStepInitial}}
\code{\link{spotStepSequential}} \code{\link{spotStepRunAlg}} \code{\link{spotStepReport}} 
\code{\link{spotGetOptions}}
}

