\name{spotWriteAroi}
\alias{spotWriteAroi}
\title{Spot Write Aroi...}
\usage{spotWriteAroi(spotConfig, aroi)}
\description{Spot Write Aroi}
\details{help function spotWriteAroi writes actual region of interest to the .aroi-file

Result/Effects: 
rewrites the actual region of interest-file}
\arguments{\item{spotConfig}{all parameters, only two are used:
\code{spotConfig$io.columnSep}: the column separator should not be empty for writing table to .des-file
\code{spotConfig$io.aroiFileName}: the filename the design should be written to}
\item{aroi}{data frame.}}
