% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/spotHelpFunctions.R
\name{spotWriteBest}
\alias{spotWriteBest}
\title{Spot Write Best}
\usage{
spotWriteBest(B, spotConfig)
}
\arguments{
\item{B}{matrix containing the merged result data of the current SPOT run}

\item{spotConfig}{all parameters, the ones of interest are:\cr
the name of the file for the best-data to be stored in: \code{spotConfig$io.bstFileName}\cr
the name string of the result column: \code{spotConfig$alg.resultColumn}\cr
the boolean that specifies if files are actually used or not: \code{spot.fileMode}\cr
the data frame that will be extended with the current best: \code{alg.currentBest}\cr}
}
\value{
returns the \code{spotConfig} list with a new row in \code{spotConfig$alg.currentBest}
}
\description{
Helper function that simply writes data to the .bst-file
(appending or creating - depends on the existence of the .bst-file)
}
\details{
Result/Effects:
adds one row to the best file and the alg.currentBest data frame.
}
\keyword{internal}

