% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{wrapBatchTools}
\alias{wrapBatchTools}
\title{wrapBatchTools}
\usage{
wrapBatchTools(fun, reg = NULL,
  clusterFunction = batchtools::makeClusterFunctionsInteractive(),
  resources = NULL)
}
\arguments{
\item{fun}{function to wrap}

\item{reg}{batchtools registry, if none is provided, then one will be created automatically}

\item{clusterFunction}{batchtools clusterFunction, default: makeClusterFunctionsInteractive()}

\item{resources}{resource list, default NULL}
}
\value{
callable function for SPOT
}
\value{
a function
}
\description{
wrapFunction to call a function through batchtools
}
